@@ -9,9 +9,11 @@ description: Demonstrates how to use the flutter_inappwebview plugin.
 # Read more about versioning at semver.org.
 version: 1.0.0+1
 
+publish_to: none
+
 environment:
-  sdk: ">=2.12.0-0 <3.0.0"
-  flutter: ">=1.22.2"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 dependencies:
   flutter:
@@ -19,11 +21,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  flutter_downloader: ^1.5.2
-  path_provider: ^2.0.0-nullsafety
-  permission_handler: ^5.1.0+2
-  url_launcher: ^6.0.0-nullsafety.4
+  cupertino_icons: ^1.0.4
+  flutter_downloader: ^1.7.3
+  path_provider: ^2.0.9
+  permission_handler: ^9.2.0
+  url_launcher: ^6.0.20
   # connectivity: ^0.4.5+6
   flutter_inappwebview:
     path: ../
