@@ -1,11 +1,11 @@
 name: flutter_inappwebview
 description: A Flutter plugin that allows you to add an inline webview, to use an headless webview, and to open an in-app browser window.
-version: 5.3.2
+version: 5.4.3+7
 homepage: https://github.com/pichillilorenzo/flutter_inappwebview
 
 environment:
-  sdk: ">=2.12.0-0 <3.0.0"
-  flutter: ">=1.22.2"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.10.0-nullsafety.1
+  pedantic: ^1.11.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
@@ -30,8 +30,8 @@ flutter:
         pluginClass: InAppWebViewFlutterPlugin
 
   assets:
-    - packages/flutter_inappwebview/t_rex_runner/t-rex.html
-    - packages/flutter_inappwebview/t_rex_runner/t-rex.css
+    - packages/flutter_inappwebview/assets/t_rex_runner/t-rex.html
+    - packages/flutter_inappwebview/assets/t_rex_runner/t-rex.css
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
   #  - images/a_dot_burr.jpeg
@@ -62,3 +62,7 @@ flutter:
   #
   # For details regarding fonts in packages, see
   # https://flutter.io/custom-fonts/#from-packages
+
+false_secrets:
+  - /nodejs_server_test_auth_basic_and_ssl/*.pem
+  - /nodejs_server_test_auth_basic_and_ssl/*.pfx
