@@ -1,13 +1,13 @@
-package com.pichillilorenzo.flutter_inappwebview.in_app_webview;
+package com.pichillilorenzo.flutter_inappwebview;
 
 import android.content.Context;
-import android.view.View;
 
-import com.pichillilorenzo.flutter_inappwebview.InAppWebViewFlutterPlugin;
+import com.pichillilorenzo.flutter_inappwebview.in_app_webview.FlutterWebView;
+import com.pichillilorenzo.flutter_inappwebview.types.PlatformWebView;
+import com.pichillilorenzo.flutter_inappwebview.types.WebViewImplementation;
 
 import java.util.HashMap;
 
-import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.StandardMessageCodec;
 import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugin.platform.PlatformViewFactory;
@@ -23,8 +23,16 @@ public class FlutterWebViewFactory extends PlatformViewFactory {
   @Override
   public PlatformView create(Context context, int id, Object args) {
     HashMap<String, Object> params = (HashMap<String, Object>) args;
-    FlutterWebView flutterWebView = new FlutterWebView(plugin, context, id, params);
+    
+    PlatformWebView flutterWebView;
+    WebViewImplementation implementation = WebViewImplementation.fromValue((Integer) params.get("implementation"));
+    switch (implementation) {
+      case NATIVE:
+      default:
+        flutterWebView = new FlutterWebView(plugin, context, id, params);
+    }
     flutterWebView.makeInitialLoad(params);
+    
     return flutterWebView;
   }
 }
