@@ -16,11 +16,15 @@ import com.pichillilorenzo.flutter_inappwebview.in_app_browser.InAppBrowserOptio
 import com.pichillilorenzo.flutter_inappwebview.in_app_webview.InAppWebView;
 import com.pichillilorenzo.flutter_inappwebview.in_app_webview.InAppWebViewOptions;
 import com.pichillilorenzo.flutter_inappwebview.types.ContentWorld;
+import com.pichillilorenzo.flutter_inappwebview.types.HitTestResult;
+import com.pichillilorenzo.flutter_inappwebview.types.InAppWebViewInterface;
 import com.pichillilorenzo.flutter_inappwebview.types.SslCertificateExt;
 import com.pichillilorenzo.flutter_inappwebview.types.URLRequest;
 import com.pichillilorenzo.flutter_inappwebview.types.UserScript;
+import com.pichillilorenzo.flutter_inappwebview.types.WebMessage;
 import com.pichillilorenzo.flutter_inappwebview.types.WebMessageChannel;
 import com.pichillilorenzo.flutter_inappwebview.types.WebMessageListener;
+import com.pichillilorenzo.flutter_inappwebview.types.WebMessagePort;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -34,9 +38,9 @@ import io.flutter.plugin.common.MethodChannel;
 public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandler {
   static final String LOG_TAG = "IAWMethodHandler";
 
-  public InAppWebView webView;
+  public InAppWebViewInterface webView;
 
-  public InAppWebViewMethodHandler(InAppWebView webView) {
+  public InAppWebViewMethodHandler(InAppWebViewInterface webView) {
     this.webView = webView;
   }
 
@@ -176,8 +180,8 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
           result.success(null);
         break;
       case "setOptions":
-        if (webView != null && webView.inAppBrowserDelegate != null && webView.inAppBrowserDelegate instanceof InAppBrowserActivity) {
-          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.inAppBrowserDelegate;
+        if (webView != null && webView.getInAppBrowserDelegate() != null && webView.getInAppBrowserDelegate() instanceof InAppBrowserActivity) {
+          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.getInAppBrowserDelegate();
           InAppBrowserOptions inAppBrowserOptions = new InAppBrowserOptions();
           HashMap<String, Object> inAppBrowserOptionsMap = (HashMap<String, Object>) call.argument("options");
           inAppBrowserOptions.parse(inAppBrowserOptionsMap);
@@ -191,24 +195,24 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         result.success(true);
         break;
       case "getOptions":
-        if (webView != null && webView.inAppBrowserDelegate != null && webView.inAppBrowserDelegate instanceof InAppBrowserActivity) {
-          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.inAppBrowserDelegate;
+        if (webView != null && webView.getInAppBrowserDelegate() != null && webView.getInAppBrowserDelegate() instanceof InAppBrowserActivity) {
+          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.getInAppBrowserDelegate();
           result.success(inAppBrowserActivity.getOptions());
         } else {
           result.success((webView != null) ? webView.getOptions() : null);
         }
         break;
       case "close":
-        if (webView != null && webView.inAppBrowserDelegate != null && webView.inAppBrowserDelegate instanceof InAppBrowserActivity) {
-          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.inAppBrowserDelegate;
+        if (webView != null && webView.getInAppBrowserDelegate() != null && webView.getInAppBrowserDelegate() instanceof InAppBrowserActivity) {
+          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.getInAppBrowserDelegate();
           inAppBrowserActivity.close(result);
         } else {
           result.notImplemented();
         }
         break;
       case "show":
-        if (webView != null && webView.inAppBrowserDelegate != null && webView.inAppBrowserDelegate instanceof InAppBrowserActivity) {
-          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.inAppBrowserDelegate;
+        if (webView != null && webView.getInAppBrowserDelegate() != null && webView.getInAppBrowserDelegate() instanceof InAppBrowserActivity) {
+          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.getInAppBrowserDelegate();
           inAppBrowserActivity.show();
           result.success(true);
         } else {
@@ -216,8 +220,8 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         }
         break;
       case "hide":
-        if (webView != null && webView.inAppBrowserDelegate != null && webView.inAppBrowserDelegate instanceof InAppBrowserActivity) {
-          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.inAppBrowserDelegate;
+        if (webView != null && webView.getInAppBrowserDelegate() != null && webView.getInAppBrowserDelegate() instanceof InAppBrowserActivity) {
+          InAppBrowserActivity inAppBrowserActivity = (InAppBrowserActivity) webView.getInAppBrowserDelegate();
           inAppBrowserActivity.hide();
           result.success(true);
         } else {
@@ -319,7 +323,11 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         result.success(true);
         break;
       case "getContentHeight":
-        result.success((webView != null) ? webView.getContentHeight() : null);
+        if (webView instanceof InAppWebView) {
+          result.success(webView.getContentHeight());
+        } else {
+          result.success(null);
+        }
         break;
       case "zoomBy":
         if (webView != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
@@ -332,22 +340,27 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         result.success((webView != null) ? webView.getOriginalUrl() : null);
         break;
       case "getZoomScale":
-        result.success((webView != null) ? webView.zoomScale : null);
+        if (webView instanceof InAppWebView) {
+          result.success(webView.getZoomScale());
+        } else {
+          result.success(null);
+        }
         break;
       case "getSelectedText":
-        if (webView != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-          webView.getSelectedText(result);
+        if ((webView instanceof InAppWebView && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT)) {
+          webView.getSelectedText(new ValueCallback<String>() {
+            @Override
+            public void onReceiveValue(String value) {
+              result.success(value);
+            }
+          });
         } else {
           result.success(null);
         }
         break;
       case "getHitTestResult":
-        if (webView != null) {
-          WebView.HitTestResult hitTestResult = webView.getHitTestResult();
-          Map<String, Object> obj = new HashMap<>();
-          obj.put("type", hitTestResult.getType());
-          obj.put("extra", hitTestResult.getExtra());
-          result.success(obj);
+        if (webView instanceof InAppWebView) {
+          result.success(HitTestResult.fromWebViewHitTestResult(webView.getHitTestResult()).toMap());
         } else {
           result.success(null);
         }
@@ -405,7 +418,7 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
       case "setContextMenu":
         if (webView != null) {
           Map<String, Object> contextMenu = (Map<String, Object>) call.argument("contextMenu");
-          webView.contextMenu = contextMenu;
+          webView.setContextMenu(contextMenu);
         }
         result.success(true);
         break;
@@ -451,34 +464,34 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         result.success(true);
         break;
       case "addUserScript":
-        if (webView != null && webView.userContentController != null) {
+        if (webView != null && webView.getUserContentController() != null) {
           Map<String, Object> userScriptMap = (Map<String, Object>) call.argument("userScript");
           UserScript userScript = UserScript.fromMap(userScriptMap);
-          result.success(webView.userContentController.addUserOnlyScript(userScript));
+          result.success(webView.getUserContentController().addUserOnlyScript(userScript));
         } else {
           result.success(false);
         }
         break;
       case "removeUserScript":
-        if (webView != null && webView.userContentController != null) {
+        if (webView != null && webView.getUserContentController() != null) {
           Integer index = (Integer) call.argument("index");
           Map<String, Object> userScriptMap = (Map<String, Object>) call.argument("userScript");
           UserScript userScript = UserScript.fromMap(userScriptMap);
-          result.success(webView.userContentController.removeUserOnlyScriptAt(index, userScript.getInjectionTime()));
+          result.success(webView.getUserContentController().removeUserOnlyScriptAt(index, userScript.getInjectionTime()));
         } else {
           result.success(false);
         }
         break;
       case "removeUserScriptsByGroupName":
-        if (webView != null && webView.userContentController != null) {
+        if (webView != null && webView.getUserContentController() != null) {
           String groupName = (String) call.argument("groupName");
-          webView.userContentController.removeUserOnlyScriptsByGroupName(groupName);
+          webView.getUserContentController().removeUserOnlyScriptsByGroupName(groupName);
         }
         result.success(true);
         break;
       case "removeAllUserScripts":
-        if (webView != null && webView.userContentController != null) {
-          webView.userContentController.removeAllUserOnlyScripts();
+        if (webView != null && webView.getUserContentController() != null) {
+          webView.getUserContentController().removeAllUserOnlyScripts();
         }
         result.success(true);
         break;
@@ -512,8 +525,12 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         }
         break;
       case "createWebMessageChannel":
-        if (webView != null && WebViewFeature.isFeatureSupported(WebViewFeature.CREATE_WEB_MESSAGE_CHANNEL)) {
-          result.success(webView.createCompatWebMessageChannel().toMap());
+        if (webView != null) {
+          if (webView instanceof InAppWebView && WebViewFeature.isFeatureSupported(WebViewFeature.CREATE_WEB_MESSAGE_CHANNEL)) {
+            result.success(webView.createCompatWebMessageChannel().toMap());
+          } else {
+            result.success(null);
+          }
         } else {
           result.success(null);
         }
@@ -522,38 +539,47 @@ public class InAppWebViewMethodHandler implements MethodChannel.MethodCallHandle
         if (webView != null && WebViewFeature.isFeatureSupported(WebViewFeature.POST_WEB_MESSAGE)) {
           Map<String, Object> message = (Map<String, Object>) call.argument("message");
           String targetOrigin = (String) call.argument("targetOrigin");
-          List<WebMessagePortCompat> ports = new ArrayList<>();
+          List<WebMessagePortCompat> compatPorts = new ArrayList<>();
+          List<WebMessagePort> ports = new ArrayList<>();
           List<Map<String, Object>> portsMap = (List<Map<String, Object>>) message.get("ports");
           if (portsMap != null) {
             for (Map<String, Object> portMap : portsMap) {
               String webMessageChannelId = (String) portMap.get("webMessageChannelId");
               Integer index = (Integer) portMap.get("index");
-              WebMessageChannel webMessageChannel = webView.webMessageChannels.get(webMessageChannelId);
+              WebMessageChannel webMessageChannel = webView.getWebMessageChannels().get(webMessageChannelId);
               if (webMessageChannel != null) {
-                ports.add(webMessageChannel.ports.get(index));
+                if (webView instanceof InAppWebView) {
+                  compatPorts.add(webMessageChannel.compatPorts.get(index));
+                }
               }
             }
           }
-          WebMessageCompat webMessage = new WebMessageCompat((String) message.get("data"), ports.toArray(new WebMessagePortCompat[0]));
-          try {
-            WebViewCompat.postWebMessage(webView, webMessage, Uri.parse(targetOrigin));
-            result.success(true);
-          } catch (Exception e) {
-            result.error(LOG_TAG, e.getMessage(), null);
+          if (webView instanceof InAppWebView) {
+            WebMessageCompat webMessage = new WebMessageCompat((String) message.get("data"), compatPorts.toArray(new WebMessagePortCompat[0]));
+            try {
+              WebViewCompat.postWebMessage((WebView) webView, webMessage, Uri.parse(targetOrigin));
+              result.success(true);
+            } catch (Exception e) {
+              result.error(LOG_TAG, e.getMessage(), null);
+            }
           }
         } else {
           result.success(true);
         }
         break;
       case "addWebMessageListener":
-        if (webView != null && WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_LISTENER)) {
+        if (webView != null) {
           Map<String, Object> webMessageListenerMap = (Map<String, Object>) call.argument("webMessageListener");
-          WebMessageListener webMessageListener = WebMessageListener.fromMap(webView.plugin.messenger, webMessageListenerMap);
-          try {
-            webView.addWebMessageListener(webMessageListener);
+          WebMessageListener webMessageListener = WebMessageListener.fromMap(webView, webView.getPlugin().messenger, webMessageListenerMap);
+          if (webView instanceof InAppWebView && WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_LISTENER)) {
+            try {
+              webView.addWebMessageListener(webMessageListener);
+              result.success(true);
+            } catch (Exception e) {
+              result.error(LOG_TAG, e.getMessage(), null);
+            }
+          } else {
             result.success(true);
-          } catch (Exception e) {
-            result.error(LOG_TAG, e.getMessage(), null);
           }
         } else {
           result.success(true);
