@@ -19,6 +19,9 @@ import org.json.JSONObject;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.Inet6Address;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 import java.security.Key;
 import java.security.KeyStore;
 import java.security.PrivateKey;
@@ -26,6 +29,7 @@ import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
+import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.List;
@@ -33,6 +37,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
+import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import javax.net.ssl.HostnameVerifier;
@@ -77,7 +82,7 @@ public class Util {
     return ANDROID_ASSET_URL + key;
   }
 
-  public static InputStream getFileAsset(InAppWebViewFlutterPlugin plugin,String assetFilePath) throws IOException {
+  public static InputStream getFileAsset(InAppWebViewFlutterPlugin plugin, String assetFilePath) throws IOException {
     String key = (plugin.registrar != null) ? plugin.registrar.lookupKeyForAsset(assetFilePath) : plugin.flutterAssets.getAssetFilePathByName(assetFilePath);
     AssetManager mg = plugin.applicationContext.getResources().getAssets();
     return mg.open(key);
@@ -286,5 +291,34 @@ public class Util {
 
   public static float getPixelDensity(Context context) {
     return context.getResources().getDisplayMetrics().density;
+  }
+
+  public static boolean isClass(String className) {
+    try  {
+      Class.forName(className);
+      return true;
+    }  catch (ClassNotFoundException e) {
+      return false;
+    }
+  }
+  
+  public static boolean isIPv6(String address) {
+    try {
+      Inet6Address.getByName(address);
+    } catch (UnknownHostException e) {
+      return false;
+    }
+    return true;
+  }
+
+  public static String normalizeIPv6(String address) throws Exception {
+    if (!Util.isIPv6(address)) {
+      throw new Exception("Invalid address: " + address);
+    }
+    return InetAddress.getByName(address).getCanonicalHostName();
+  }
+
+  public static Object getOrDefault(Map map, String key, Object defaultValue) {
+    return map.containsKey(key) ? map.get(key) : defaultValue;
   }
 }
