@@ -1,6 +1,7 @@
 package com.pichillilorenzo.flutter_inappwebview.chrome_custom_tabs;
 
 import android.app.Activity;
+import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
 
@@ -9,6 +10,7 @@ import androidx.browser.customtabs.CustomTabsClient;
 import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.browser.customtabs.CustomTabsServiceConnection;
 import androidx.browser.customtabs.CustomTabsSession;
+import androidx.browser.trusted.TrustedWebActivityIntent;
 
 import java.util.List;
 
@@ -37,6 +39,14 @@ public class CustomTabActivityHelper implements ServiceConnectionCallback {
         activity.startActivityForResult(customTabsIntent.intent, requestCode);
     }
 
+    public static void openCustomTab(Activity activity,
+                                     TrustedWebActivityIntent trustedWebActivityIntent,
+                                     Uri uri,
+                                     int requestCode) {
+        trustedWebActivityIntent.getIntent().setData(uri);
+        activity.startActivityForResult(trustedWebActivityIntent.getIntent(), requestCode);
+    }
+    
     public static boolean isAvailable(Activity activity) {
         return CustomTabsHelper.getPackageNameToUse(activity) != null;
     }
@@ -135,5 +145,4 @@ public class CustomTabActivityHelper implements ServiceConnectionCallback {
          */
         void onCustomTabsDisconnected();
     }
-
 }
