@@ -20,6 +20,7 @@ import android.webkit.WebViewClient;
 import android.widget.ProgressBar;
 import android.widget.SearchView;
 
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -48,17 +49,26 @@ public class InAppBrowserActivity extends AppCompatActivity implements InAppBrow
   public MethodChannel channel;
   public Integer windowId;
   public String id;
+  @Nullable
   public InAppWebView webView;
+  @Nullable
   public PullToRefreshLayout pullToRefreshLayout;
+  @Nullable
   public ActionBar actionBar;
+  @Nullable
   public Menu menu;
+  @Nullable
   public SearchView searchView;
-  public InAppBrowserOptions options;
+  public InAppBrowserOptions options = new InAppBrowserOptions();
+  @Nullable
   public ProgressBar progressBar;
   public boolean isHidden = false;
+  @Nullable
   public String fromActivity;
   private List<ActivityResultListener> activityResultListeners = new ArrayList<>();
+  @Nullable
   public InAppWebViewMethodHandler methodCallDelegate;
+  @Nullable
   public InAppBrowserManager manager;
   
   @Override
@@ -66,12 +76,17 @@ public class InAppBrowserActivity extends AppCompatActivity implements InAppBrow
     super.onCreate(savedInstanceState);
 
     Bundle b = getIntent().getExtras();
-    assert b != null;
+    if (b == null) return;
+    
     id = b.getString("id");
 
     String managerId = b.getString("managerId");
-    manager = (InAppBrowserManager) InAppBrowserManager.shared.get(managerId);
-    
+    manager = InAppBrowserManager.shared.get(managerId);
+    if (manager == null || manager.plugin == null|| manager.plugin.messenger == null) return;
+
+    Map<String, Object> optionsMap = (Map<String, Object>) b.getSerializable("options");
+    options.parse(optionsMap);
+
     windowId = b.getInt("windowId");
 
     channel = new MethodChannel(manager.plugin.messenger, "com.pichillilorenzo/flutter_inappbrowser_" + id);
@@ -98,12 +113,8 @@ public class InAppBrowserActivity extends AppCompatActivity implements InAppBrow
 
     fromActivity = b.getString("fromActivity");
 
-    Map<String, Object> optionsMap = (Map<String, Object>) b.getSerializable("options");
     Map<String, Object> contextMenu = (Map<String, Object>) b.getSerializable("contextMenu");
     List<Map<String, Object>> initialUserScripts = (List<Map<String, Object>>) b.getSerializable("initialUserScripts");
-
-    options = new InAppBrowserOptions();
-    options.parse(optionsMap);
 
     InAppWebViewOptions webViewOptions = new InAppWebViewOptions();
     webViewOptions.parse(optionsMap);
@@ -178,75 +189,87 @@ public class InAppBrowserActivity extends AppCompatActivity implements InAppBrow
     else
       progressBar.setMax(100);
 
-    actionBar.setDisplayShowTitleEnabled(!options.hideTitleBar);
+    if (actionBar != null) {
+      actionBar.setDisplayShowTitleEnabled(!options.hideTitleBar);
 
-    if (options.hideToolbarTop)
-      actionBar.hide();
+      if (options.hideToolbarTop)
+        actionBar.hide();
 
-    if (options.toolbarTopBackgroundColor != null && !options.toolbarTopBackgroundColor.isEmpty())
-      actionBar.setBackgroundDrawable(new ColorDrawable(Color.parseColor(options.toolbarTopBackgroundColor)));
+      if (options.toolbarTopBackgroundColor != null && !options.toolbarTopBackgroundColor.isEmpty())
+        actionBar.setBackgroundDrawable(new ColorDrawable(Color.parseColor(options.toolbarTopBackgroundColor)));
 
-    if (options.toolbarTopFixedTitle != null && !options.toolbarTopFixedTitle.isEmpty())
-      actionBar.setTitle(options.toolbarTopFixedTitle);
-
+      if (options.toolbarTopFixedTitle != null && !options.toolbarTopFixedTitle.isEmpty())
+        actionBar.setTitle(options.toolbarTopFixedTitle);
+    }
   }
 
   @Override
   public boolean onCreateOptionsMenu(Menu m) {
     menu = m;
 
+    if (actionBar != null && (options.toolbarTopFixedTitle == null || options.toolbarTopFixedTitle.isEmpty()))
+      actionBar.setTitle(webView != null ? webView.getTitle() : "");
+
+    if (menu == null)
+      return super.onCreateOptionsMenu(m);
+
     MenuInflater inflater = getMenuInflater();
     // Inflate menu to add items to action bar if it is present.
     inflater.inflate(R.menu.menu_main, menu);
 
-    searchView = (SearchView) menu.findItem(R.id.menu_search).getActionView();
-    searchView.setFocusable(true);
+    MenuItem menuItem = menu.findItem(R.id.menu_search);
+    if (menuItem != null) {
+      if (options.hideUrlBar)
+        menuItem.setVisible(false);
 
-    if (options.hideUrlBar)
-      menu.findItem(R.id.menu_search).setVisible(false);
+      searchView = (SearchView) menuItem.getActionView();
+      if (searchView != null) {
+        searchView.setFocusable(true);
 
-    searchView.setQuery(webView.getUrl(), false);
+        searchView.setQuery(webView != null ? webView.getUrl() : "", false);
 
-    if (options.toolbarTopFixedTitle == null || options.toolbarTopFixedTitle.isEmpty())
-      actionBar.setTitle(webView.getTitle());
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+          @Override
+          public boolean onQueryTextSubmit(String query) {
+            if (!query.isEmpty()) {
+              if (webView != null)
+                webView.loadUrl(query);
+              if (searchView != null) {
+                searchView.setQuery("", false);
+                searchView.setIconified(true);
+              }
+              return true;
+            }
+            return false;
+          }
 
-    searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-      @Override
-      public boolean onQueryTextSubmit(String query) {
-        if (!query.isEmpty()) {
-          webView.loadUrl(query);
-          searchView.setQuery("", false);
-          searchView.setIconified(true);
-          return true;
-        }
-        return false;
-      }
+          @Override
+          public boolean onQueryTextChange(String newText) {
+            return false;
+          }
 
-      @Override
-      public boolean onQueryTextChange(String newText) {
-        return false;
-      }
+        });
 
-    });
+        searchView.setOnCloseListener(new SearchView.OnCloseListener() {
+          @Override
+          public boolean onClose() {
+            if (searchView != null && searchView.getQuery().toString().isEmpty())
+              searchView.setQuery(webView != null ? webView.getUrl() : "", false);
+            return false;
+          }
+        });
 
-    searchView.setOnCloseListener(new SearchView.OnCloseListener() {
-      @Override
-      public boolean onClose() {
-        if (searchView.getQuery().toString().isEmpty())
-          searchView.setQuery(webView.getUrl(), false);
-        return false;
-      }
-    });
-
-    searchView.setOnQueryTextFocusChangeListener(new View.OnFocusChangeListener() {
-      @Override
-      public void onFocusChange(View view, boolean b) {
-        if (!b) {
-          searchView.setQuery("", false);
-          searchView.setIconified(true);
-        }
-      }
-    });
+        searchView.setOnQueryTextFocusChangeListener(new View.OnFocusChangeListener() {
+          @Override
+          public void onFocusChange(View view, boolean b) {
+            if (!b && searchView != null) {
+              searchView.setQuery("", false);
+              searchView.setIconified(true);
+            }
+          }
+        });
+      } 
+    }
 
     return true;
   }
@@ -371,21 +394,24 @@ public class InAppBrowserActivity extends AppCompatActivity implements InAppBrow
         progressBar.setMax(100);
     }
 
-    if (newOptionsMap.get("hideTitleBar") != null && options.hideTitleBar != newOptions.hideTitleBar)
+    if (actionBar != null && newOptionsMap.get("hideTitleBar") != null && options.hideTitleBar != newOptions.hideTitleBar)
       actionBar.setDisplayShowTitleEnabled(!newOptions.hideTitleBar);
 
-    if (newOptionsMap.get("hideToolbarTop") != null && options.hideToolbarTop != newOptions.hideToolbarTop) {
+    if (actionBar != null && newOptionsMap.get("hideToolbarTop") != null && options.hideToolbarTop != newOptions.hideToolbarTop) {
       if (newOptions.hideToolbarTop)
         actionBar.hide();
       else
         actionBar.show();
     }
 
-    if (newOptionsMap.get("toolbarTopBackgroundColor") != null && !Util.objEquals(options.toolbarTopBackgroundColor, newOptions.toolbarTopBackgroundColor) &&
+    if (actionBar != null && newOptionsMap.get("toolbarTopBackgroundColor") != null &&
+            !Util.objEquals(options.toolbarTopBackgroundColor, newOptions.toolbarTopBackgroundColor) &&
             !newOptions.toolbarTopBackgroundColor.isEmpty())
       actionBar.setBackgroundDrawable(new ColorDrawable(Color.parseColor(newOptions.toolbarTopBackgroundColor)));
 
-    if (newOptionsMap.get("toolbarTopFixedTitle") != null && !Util.objEquals(options.toolbarTopFixedTitle, newOptions.toolbarTopFixedTitle) && !newOptions.toolbarTopFixedTitle.isEmpty())
+    if (actionBar != null && newOptionsMap.get("toolbarTopFixedTitle") != null &&
+            !Util.objEquals(options.toolbarTopFixedTitle, newOptions.toolbarTopFixedTitle) &&
+            !newOptions.toolbarTopFixedTitle.isEmpty())
       actionBar.setTitle(newOptions.toolbarTopFixedTitle);
 
     if (newOptionsMap.get("hideUrlBar") != null && options.hideUrlBar != newOptions.hideUrlBar) {
@@ -493,7 +519,7 @@ public class InAppBrowserActivity extends AppCompatActivity implements InAppBrow
       methodCallDelegate = null;
     }
     if (webView != null) {
-      if (manager.plugin.activityPluginBinding != null) {
+      if (manager != null && manager.plugin != null && manager.plugin.activityPluginBinding != null) {
         manager.plugin.activityPluginBinding.removeActivityResultListener(webView.inAppWebViewChromeClient);
       }
       ViewGroup vg = (ViewGroup) (webView.getParent());
