@@ -11,6 +11,7 @@ import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Build;
+import android.os.Bundle;
 import android.os.Environment;
 import android.os.Message;
 import android.os.Parcelable;
@@ -112,7 +113,7 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
 
     if (plugin.registrar != null)
       plugin.registrar.addActivityResultListener(this);
-    else
+    else if (plugin.activityPluginBinding != null)
       plugin.activityPluginBinding.addActivityResultListener(this);
   }
 
@@ -123,9 +124,15 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
 
   @Override
   public void onHideCustomView() {
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return;
+    }
 
     View decorView = getRootView();
+    if (decorView == null) {
+      return;
+    }
     ((FrameLayout) decorView).removeView(this.mCustomView);
     this.mCustomView = null;
     decorView.setSystemUiVisibility(this.mOriginalSystemUiVisibility);
@@ -144,9 +151,15 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
       return;
     }
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return;
+    }
 
     View decorView = getRootView();
+    if (decorView == null) {
+      return;
+    }
     this.mCustomView = paramView;
     this.mOriginalSystemUiVisibility = decorView.getSystemUiVisibility();
     this.mOriginalOrientation = activity.getRequestedOrientation();
@@ -228,7 +241,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
       }
     };
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return;
+    }
 
     AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(activity, R.style.Theme_AppCompat_Dialog_Alert);
     alertDialogBuilder.setMessage(alertMessage);
@@ -321,7 +337,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
       }
     };
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return;
+    }
 
     AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(activity, R.style.Theme_AppCompat_Dialog_Alert);
     alertDialogBuilder.setMessage(alertMessage);
@@ -440,7 +459,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
       }
     };
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return;
+    }
 
     AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(activity, R.style.Theme_AppCompat_Dialog_Alert);
     alertDialogBuilder.setMessage(alertMessage);
@@ -538,7 +560,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
         }
       };
 
-      Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+      Activity activity = getActivity();
+      if (activity == null) {
+        return;
+      }
 
       AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(activity, R.style.Theme_AppCompat_Dialog_Alert);
       alertDialogBuilder.setMessage(alertMessage);
@@ -572,6 +597,19 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
 
     WebView.HitTestResult result = view.getHitTestResult();
     String url = result.getExtra();
+
+    // Ensure that images with hyperlink return the correct URL, not the image source
+    if(result.getType() == WebView.HitTestResult.SRC_IMAGE_ANCHOR_TYPE) {
+      Message href = view.getHandler().obtainMessage();
+      view.requestFocusNodeHref(href);
+      Bundle data = href.getData();
+      if (data != null) {
+        String imageUrl = data.getString("url");
+        if(imageUrl != null && !imageUrl.isEmpty()) {
+          url = imageUrl;
+        }
+      }
+    }
 
     URLRequest request = new URLRequest(url, "GET", null, null);
     CreateWindowAction createWindowAction = new CreateWindowAction(
@@ -732,8 +770,12 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
     channel.invokeMethod("onReceivedTouchIconUrl", obj);
   }
 
+  @Nullable
   protected ViewGroup getRootView() {
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return null;
+    }
     return (ViewGroup) activity.findViewById(android.R.id.content);
   }
 
@@ -827,7 +869,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
   }
 
   private boolean isFileNotEmpty(Uri uri) {
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return false;
+    }
 
     long length;
     try {
@@ -868,7 +913,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
     }
     chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, extraIntents.toArray(new Parcelable[]{}));
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return;
+    }
     if (chooserIntent.resolveActivity(activity.getPackageManager()) != null) {
       activity.startActivityForResult(chooserIntent, PICKER_LEGACY);
     } else {
@@ -896,7 +944,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
     chooserIntent.putExtra(Intent.EXTRA_INTENT, fileSelectionIntent);
     chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, extraIntents.toArray(new Parcelable[]{}));
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return true;
+    }
     if (chooserIntent.resolveActivity(activity.getPackageManager()) != null) {
       activity.startActivityForResult(chooserIntent, PICKER);
     } else {
@@ -909,7 +960,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
   protected boolean needsCameraPermission() {
     boolean needed = false;
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return true;
+    }
     PackageManager packageManager = activity.getPackageManager();
     try {
       String[] requestedPermissions = packageManager.getPackageInfo(activity.getApplicationContext().getPackageName(), PackageManager.GET_PERMISSIONS).requestedPermissions;
@@ -1004,7 +1058,7 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
 
   private Boolean arrayContainsString(String[] array, String pattern) {
     for (String content : array) {
-      if (content.contains(pattern)) {
+      if (content != null && content.contains(pattern)) {
         return true;
       }
     }
@@ -1037,6 +1091,7 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
     return type;
   }
 
+  @Nullable
   private Uri getOutputUri(String intentType) {
     File capturedFile = null;
     try {
@@ -1051,12 +1106,16 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
       return Uri.fromFile(capturedFile);
     }
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return null;
+    }
     // for versions 6.0+ (23) we use the FileProvider to avoid runtime permissions
     String packageName = activity.getApplicationContext().getPackageName();
     return FileProvider.getUriForFile(activity.getApplicationContext(), packageName + "." + fileProviderAuthorityExtension, capturedFile);
   }
 
+  @Nullable
   private File getCapturedFile(String intentType) throws IOException {
     String prefix = "";
     String suffix = "";
@@ -1081,7 +1140,10 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
       return new File(storageDir, filename);
     }
 
-    Activity activity = inAppBrowserDelegate != null ? inAppBrowserDelegate.getActivity() : plugin.activity;
+    Activity activity = getActivity();
+    if (activity == null) {
+      return null;
+    }
     File storageDir = activity.getApplicationContext().getExternalFilesDir(null);
     return File.createTempFile(prefix, suffix, storageDir);
   }
@@ -1137,6 +1199,16 @@ public class InAppWebViewChromeClient extends WebChromeClient implements PluginR
         }
       });
     }
+  }
+
+  @Nullable
+  private Activity getActivity() {
+    if (inAppBrowserDelegate != null) {
+      return inAppBrowserDelegate.getActivity();
+    } else if (plugin != null) {
+      return plugin.activity;
+    }
+    return null;
   }
 
   public void dispose() {
