@@ -1,6 +1,5 @@
 package com.pichillilorenzo.flutter_inappwebview.types;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import java.util.Arrays;
@@ -8,7 +7,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class URLRequest {
-  @NonNull
+  @Nullable
   private String url;
   @Nullable
   private String method;
@@ -17,7 +16,7 @@ public class URLRequest {
   @Nullable
   private Map<String, String> headers;
 
-  public URLRequest(@NonNull String url, @Nullable String method, @Nullable byte[] body, @Nullable Map<String, String> headers) {
+  public URLRequest(@Nullable String url, @Nullable String method, @Nullable byte[] body, @Nullable Map<String, String> headers) {
     this.url = url;
     this.method = method;
     this.body = body;
@@ -30,10 +29,12 @@ public class URLRequest {
       return null;
     }
     String url = (String) map.get("url");
+    if (url == null) {
+      url = "about:blank";
+    }
     String method = (String) map.get("method");
     byte[] body = (byte[]) map.get("body");
     Map<String, String> headers = (Map<String, String>) map.get("headers");
-    assert url != null;
     return new URLRequest(url, method, body, headers);
   }
 
@@ -45,12 +46,12 @@ public class URLRequest {
     return urlRequestMap;
   }
 
-  @NonNull
+  @Nullable
   public String getUrl() {
     return url;
   }
 
-  public void setUrl(@NonNull String url) {
+  public void setUrl(@Nullable String url) {
     this.url = url;
   }
 
@@ -88,7 +89,7 @@ public class URLRequest {
 
     URLRequest that = (URLRequest) o;
 
-    if (!url.equals(that.url)) return false;
+    if (url != null ? !url.equals(that.url) : that.url != null) return false;
     if (method != null ? !method.equals(that.method) : that.method != null) return false;
     if (!Arrays.equals(body, that.body)) return false;
     return headers != null ? headers.equals(that.headers) : that.headers == null;
@@ -96,7 +97,7 @@ public class URLRequest {
 
   @Override
   public int hashCode() {
-    int result = url.hashCode();
+    int result = url != null ? url.hashCode() : 0;
     result = 31 * result + (method != null ? method.hashCode() : 0);
     result = 31 * result + Arrays.hashCode(body);
     result = 31 * result + (headers != null ? headers.hashCode() : 0);
