@@ -1,5 +1,7 @@
 package com.pichillilorenzo.flutter_inappwebview.types;
 
+import android.webkit.ValueCallback;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.webkit.WebMessageCompat;
@@ -8,6 +10,7 @@ import androidx.webkit.WebViewCompat;
 import androidx.webkit.WebViewFeature;
 
 import com.pichillilorenzo.flutter_inappwebview.in_app_webview.InAppWebView;
+import com.pichillilorenzo.flutter_inappwebview.plugin_scripts_js.JavaScriptBridgeJS;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -23,85 +26,64 @@ public class WebMessageChannel implements MethodChannel.MethodCallHandler {
 
   public String id;
   public MethodChannel channel;
-  public final List<WebMessagePortCompat> ports;
-  private InAppWebView webView;
+  public final List<WebMessagePortCompat> compatPorts;
+  public final List<WebMessagePort> ports;
+  public InAppWebViewInterface webView;
 
-  public WebMessageChannel(@NonNull String id, @NonNull InAppWebView webView) {
+  public WebMessageChannel(@NonNull String id, @NonNull InAppWebViewInterface webView) {
     this.id = id;
-    this.channel = new MethodChannel(webView.plugin.messenger, "com.pichillilorenzo/flutter_inappwebview_web_message_channel_" + id);
+    this.channel = new MethodChannel(webView.getPlugin().messenger, "com.pichillilorenzo/flutter_inappwebview_web_message_channel_" + id);
     this.channel.setMethodCallHandler(this);
-    this.ports = new ArrayList<>(Arrays.asList(WebViewCompat.createWebMessageChannel(webView)));
+    if (webView instanceof InAppWebView) {
+      this.compatPorts = new ArrayList<>(Arrays.asList(WebViewCompat.createWebMessageChannel((InAppWebView) webView)));
+      this.ports = new ArrayList<>();
+    } else {
+      this.ports = Arrays.asList(new WebMessagePort("port1", this), new WebMessagePort("port2", this));
+      this.compatPorts = new ArrayList<>();
+    }
     this.webView = webView;
+  }
+
+  public void initJsInstance(InAppWebViewInterface webView, final ValueCallback<WebMessageChannel> callback) {
+    if (webView != null) {
+      final WebMessageChannel webMessageChannel = this;
+      webView.evaluateJavascript("(function() {" +
+              JavaScriptBridgeJS.WEB_MESSAGE_CHANNELS_VARIABLE_NAME + "['" + webMessageChannel.id + "'] = new MessageChannel();" +
+              "})();", null, new ValueCallback<String>() {
+        @Override
+        public void onReceiveValue(String value) {
+          callback.onReceiveValue(webMessageChannel);
+        }
+      });
+    } else {
+      callback.onReceiveValue(this);
+    }
   }
 
   @Override
   public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
     switch (call.method) {
       case "setWebMessageCallback":
-        if (webView != null && ports.size() > 0 &&
-                WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_SET_MESSAGE_CALLBACK)) {
+        if (webView instanceof InAppWebView) {
           final Integer index = (Integer) call.argument("index");
-          final WebMessagePortCompat webMessagePort = ports.get(index);
-          try {
-            webMessagePort.setWebMessageCallback(new WebMessagePortCompat.WebMessageCallbackCompat() {
-              @Override
-              public void onMessage(@NonNull WebMessagePortCompat port, @Nullable WebMessageCompat message) {
-                super.onMessage(port, message);
-
-                Map<String, Object> obj = new HashMap<>();
-                obj.put("index", index);
-                obj.put("message", message != null ? message.getData() : null);
-                channel.invokeMethod("onMessage", obj);
-              }
-            });
-            result.success(true);
-          } catch (Exception e) {
-            result.error(LOG_TAG, e.getMessage(), null);
-          }
+          setWebMessageCallbackForInAppWebView(index, result);
         } else {
           result.success(true);
         }
         break;
       case "postMessage":
-        if (webView != null && ports.size() > 0 &&
-                WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_POST_MESSAGE)) {
+        if (webView instanceof InAppWebView) {
           final Integer index = (Integer) call.argument("index");
-          WebMessagePortCompat port = ports.get(index);
           Map<String, Object> message = (Map<String, Object>) call.argument("message");
-          List<WebMessagePortCompat> webMessagePorts = new ArrayList<>();
-          List<Map<String, Object>> portsMap = (List<Map<String, Object>>) message.get("ports");
-          if (portsMap != null) {
-            for (Map<String, Object> portMap : portsMap) {
-              String webMessageChannelId = (String) portMap.get("webMessageChannelId");
-              Integer portIndex = (Integer) portMap.get("index");
-              WebMessageChannel webMessageChannel = webView.webMessageChannels.get(webMessageChannelId);
-              if (webMessageChannel != null) {
-                webMessagePorts.add(webMessageChannel.ports.get(portIndex));
-              }
-            }
-          }
-          WebMessageCompat webMessage = new WebMessageCompat((String) message.get("data"), webMessagePorts.toArray(new WebMessagePortCompat[0]));
-          try {
-            port.postMessage(webMessage);
-            result.success(true);
-          } catch (Exception e) {
-            result.error(LOG_TAG, e.getMessage(), null);
-          }
+          postMessageForInAppWebView(index, message, result);
         } else {
           result.success(true);
         }
         break;
       case "close":
-        if (webView != null && ports.size() > 0 &&
-                WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_CLOSE)) {
+        if (webView instanceof InAppWebView) {
           Integer index = (Integer) call.argument("index");
-          WebMessagePortCompat port = ports.get(index);
-          try {
-            port.close();
-            result.success(true);
-          } catch (Exception e) {
-            result.error(LOG_TAG, e.getMessage(), null);
-          }
+          closeForInAppWebView(index, result);
         } else {
           result.success(true);
         }
@@ -109,6 +91,81 @@ public class WebMessageChannel implements MethodChannel.MethodCallHandler {
       default:
         result.notImplemented();
     }
+  }
+  
+  private void setWebMessageCallbackForInAppWebView(final Integer index, @NonNull MethodChannel.Result result) {
+    if (webView != null && compatPorts.size() > 0 &&
+            WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_SET_MESSAGE_CALLBACK)) {
+      final WebMessagePortCompat webMessagePort = compatPorts.get(index);
+      try {
+        webMessagePort.setWebMessageCallback(new WebMessagePortCompat.WebMessageCallbackCompat() {
+          @Override
+          public void onMessage(@NonNull WebMessagePortCompat port, @Nullable WebMessageCompat message) {
+            super.onMessage(port, message);
+
+            Map<String, Object> obj = new HashMap<>();
+            obj.put("index", index);
+            obj.put("message", message != null ? message.getData() : null);
+            channel.invokeMethod("onMessage", obj);
+          }
+        });
+        result.success(true);
+      } catch (Exception e) {
+        result.error(LOG_TAG, e.getMessage(), null);
+      }
+    } else {
+      result.success(true);
+    }
+  }
+
+  private void postMessageForInAppWebView(final Integer index, Map<String, Object> message, @NonNull MethodChannel.Result result) {
+    if (webView != null && compatPorts.size() > 0 &&
+            WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_POST_MESSAGE)) {
+      WebMessagePortCompat port = compatPorts.get(index);
+      List<WebMessagePortCompat> webMessagePorts = new ArrayList<>();
+      List<Map<String, Object>> portsMap = (List<Map<String, Object>>) message.get("ports");
+      if (portsMap != null) {
+        for (Map<String, Object> portMap : portsMap) {
+          String webMessageChannelId = (String) portMap.get("webMessageChannelId");
+          Integer portIndex = (Integer) portMap.get("index");
+          WebMessageChannel webMessageChannel = webView.getWebMessageChannels().get(webMessageChannelId);
+          if (webMessageChannel != null) {
+            webMessagePorts.add(webMessageChannel.compatPorts.get(portIndex));
+          }
+        }
+      }
+      WebMessageCompat webMessage = new WebMessageCompat((String) message.get("data"), webMessagePorts.toArray(new WebMessagePortCompat[0]));
+      try {
+        port.postMessage(webMessage);
+        result.success(true);
+      } catch (Exception e) {
+        result.error(LOG_TAG, e.getMessage(), null);
+      }
+    } else {
+      result.success(true);
+    }
+  }
+
+  private void closeForInAppWebView(Integer index, @NonNull MethodChannel.Result result) {
+    if (webView != null && compatPorts.size() > 0 &&
+            WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_CLOSE)) {
+      WebMessagePortCompat port = compatPorts.get(index);
+      try {
+        port.close();
+        result.success(true);
+      } catch (Exception e) {
+        result.error(LOG_TAG, e.getMessage(), null);
+      }
+    } else {
+      result.success(true);
+    }
+  }
+
+  public void onMessage(Integer index, String message) {
+    Map<String, Object> obj = new HashMap<>();
+    obj.put("index", index);
+    obj.put("message", message );
+    channel.invokeMethod("onMessage", obj);
   }
 
   public Map<String, Object> toMap() {
@@ -119,14 +176,14 @@ public class WebMessageChannel implements MethodChannel.MethodCallHandler {
 
   public void dispose() {
     if (WebViewFeature.isFeatureSupported(WebViewFeature.WEB_MESSAGE_PORT_CLOSE)) {
-      for (WebMessagePortCompat port : ports) {
+      for (WebMessagePortCompat port : compatPorts) {
         try {
           port.close();
         } catch (Exception ignored) {}
       }
     }
     this.channel.setMethodCallHandler(null);
-    this.ports.clear();
+    this.compatPorts.clear();
     this.webView = null;
   }
 }
