@@ -15,6 +15,21 @@ import 'package:path_provider/path_provider.dart';
 
 import '.env.dart';
 
+/// Returns a matcher that matches the isNullOrEmpty property.
+const Matcher isNullOrEmpty = _NullOrEmpty();
+
+class _NullOrEmpty extends Matcher {
+  const _NullOrEmpty();
+
+  @override
+  bool matches(Object? item, Map matchState) =>
+      item == null || (item as dynamic).isEmpty;
+
+  @override
+  Description describe(Description description) =>
+      description.add('null or empty');
+}
+
 class Foo {
   String? bar;
   String? baz;
@@ -77,7 +92,7 @@ void main() {
     AndroidInAppWebViewController.setWebContentsDebuggingEnabled(true);
   }
 
- group('InAppWebView', () {
+  group('InAppWebView', () {
     testWidgets('initialUrlRequest', (WidgetTester tester) async {
       final Completer controllerCompleter = Completer<InAppWebViewController>();
       await tester.pumpWidget(
@@ -397,7 +412,7 @@ void main() {
       late File fileJs;
 
       setUpAll(() async {
-        appSupportDir = (await getApplicationSupportDirectory())!;
+        appSupportDir = (await getApplicationSupportDirectory());
 
         final Directory htmlFolder = Directory('${appSupportDir.path}/html/');
         if (!await htmlFolder.exists()) {
@@ -476,7 +491,7 @@ void main() {
       }, skip: !Platform.isIOS);
 
       testWidgets(
-          'loadUrl with file:// scheme and iosAllowingReadAccessTo argument',
+          'loadUrl with file:// scheme and allowingReadAccessTo argument',
           (WidgetTester tester) async {
         final Completer<ConsoleMessage?> consoleMessageShouldNotComplete =
             Completer<ConsoleMessage?>();
@@ -511,7 +526,7 @@ void main() {
                 controller.loadUrl(
                     urlRequest:
                         URLRequest(url: Uri.parse('file://${fileHtml.path}')),
-                    iosAllowingReadAccessTo:
+                    allowingReadAccessTo:
                         Uri.parse('file://${appSupportDir.path}/'));
               },
               onConsoleMessage: (controller, consoleMessage) {
@@ -1558,6 +1573,7 @@ void main() {
             onWebViewCreated: (controller) {
               controllerCompleter.complete(controller);
             },
+            initialUrlRequest: URLRequest(url: Uri.parse("about:blank")),
             initialOptions: InAppWebViewGroupOptions(
               crossPlatform: InAppWebViewOptions(
                   javaScriptEnabled: true,
@@ -1708,13 +1724,14 @@ void main() {
 
     group('intercept ajax request', () {
       testWidgets('send string data', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer shouldInterceptAjaxPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         final Completer<Map<String, dynamic>> onAjaxReadyStateChangeCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<Map<String, dynamic>> onAjaxProgressCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -1744,9 +1761,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptAjaxRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptAjaxRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
               },
@@ -1779,9 +1796,9 @@ void main() {
 
         await shouldInterceptAjaxPostRequestCompleter.future;
         final Map<String, dynamic> onAjaxReadyStateChangeValue =
-        await onAjaxReadyStateChangeCompleter.future;
+            await onAjaxReadyStateChangeCompleter.future;
         final Map<String, dynamic> onAjaxProgressValue =
-        await onAjaxProgressCompleter.future;
+            await onAjaxProgressCompleter.future;
 
         expect(
             mapEquals(onAjaxReadyStateChangeValue,
@@ -1794,13 +1811,14 @@ void main() {
       });
 
       testWidgets('send json data', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer shouldInterceptAjaxPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         final Completer<Map<String, dynamic>> onAjaxReadyStateChangeCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<Map<String, dynamic>> onAjaxProgressCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -1834,15 +1852,16 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptAjaxRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptAjaxRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
               },
               shouldInterceptAjaxRequest: (controller, ajaxRequest) async {
                 String data = ajaxRequest.data;
-                assert(data.contains('"firstname":"Foo"') && data.contains('"lastname":"Bar"'));
+                assert(data.contains('"firstname":"Foo"') &&
+                    data.contains('"lastname":"Bar"'));
 
                 ajaxRequest.responseType = 'json';
                 ajaxRequest.data = '{"firstname": "Foo2", "lastname": "Bar2"}';
@@ -1870,9 +1889,9 @@ void main() {
 
         await shouldInterceptAjaxPostRequestCompleter.future;
         final Map<String, dynamic> onAjaxReadyStateChangeValue =
-        await onAjaxReadyStateChangeCompleter.future;
+            await onAjaxReadyStateChangeCompleter.future;
         final Map<String, dynamic> onAjaxProgressValue =
-        await onAjaxProgressCompleter.future;
+            await onAjaxProgressCompleter.future;
 
         expect(
             mapEquals(onAjaxReadyStateChangeValue,
@@ -1885,13 +1904,14 @@ void main() {
       });
 
       testWidgets('send URLSearchParams data', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer shouldInterceptAjaxPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         final Completer<Map<String, dynamic>> onAjaxReadyStateChangeCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<Map<String, dynamic>> onAjaxProgressCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -1923,9 +1943,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptAjaxRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptAjaxRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
               },
@@ -1958,9 +1978,9 @@ void main() {
 
         await shouldInterceptAjaxPostRequestCompleter.future;
         final Map<String, dynamic> onAjaxReadyStateChangeValue =
-        await onAjaxReadyStateChangeCompleter.future;
+            await onAjaxReadyStateChangeCompleter.future;
         final Map<String, dynamic> onAjaxProgressValue =
-        await onAjaxProgressCompleter.future;
+            await onAjaxProgressCompleter.future;
 
         expect(
             mapEquals(onAjaxReadyStateChangeValue,
@@ -1973,13 +1993,14 @@ void main() {
       });
 
       testWidgets('send FormData', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer shouldInterceptAjaxPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         final Completer<Map<String, dynamic>> onAjaxReadyStateChangeCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<Map<String, dynamic>> onAjaxProgressCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -2011,9 +2032,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptAjaxRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptAjaxRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
               },
@@ -2024,7 +2045,9 @@ void main() {
                 var bodyString = String.fromCharCodes(body);
                 assert(bodyString.indexOf("WebKitFormBoundary") >= 0);
 
-                ajaxRequest.data = utf8.encode(bodyString.replaceFirst("Foo", "Foo2").replaceFirst("Bar", "Bar2"));
+                ajaxRequest.data = utf8.encode(bodyString
+                    .replaceFirst("Foo", "Foo2")
+                    .replaceFirst("Bar", "Bar2"));
                 ajaxRequest.responseType = 'json';
                 shouldInterceptAjaxPostRequestCompleter.complete(controller);
                 return ajaxRequest;
@@ -2050,9 +2073,9 @@ void main() {
 
         await shouldInterceptAjaxPostRequestCompleter.future;
         final Map<String, dynamic> onAjaxReadyStateChangeValue =
-        await onAjaxReadyStateChangeCompleter.future;
+            await onAjaxReadyStateChangeCompleter.future;
         final Map<String, dynamic> onAjaxProgressValue =
-        await onAjaxProgressCompleter.future;
+            await onAjaxProgressCompleter.future;
 
         expect(
             mapEquals(onAjaxReadyStateChangeValue,
@@ -2067,11 +2090,12 @@ void main() {
 
     group('intercept fetch request', () {
       testWidgets('send string data', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer<Map<String, dynamic>> fetchPostCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<void> shouldInterceptFetchPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -2112,9 +2136,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptFetchRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptFetchRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
 
@@ -2146,11 +2170,12 @@ void main() {
       });
 
       testWidgets('send json data', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer<Map<String, dynamic>> fetchPostCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<void> shouldInterceptFetchPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -2195,9 +2220,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptFetchRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptFetchRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
 
@@ -2210,7 +2235,8 @@ void main() {
               },
               shouldInterceptFetchRequest: (controller, fetchRequest) async {
                 String body = fetchRequest.body;
-                assert(body.contains('"firstname":"Foo"') && body.contains('"lastname":"Bar"'));
+                assert(body.contains('"firstname":"Foo"') &&
+                    body.contains('"lastname":"Bar"'));
 
                 fetchRequest.body = '{"firstname": "Foo2", "lastname": "Bar2"}';
                 shouldInterceptFetchPostRequestCompleter.complete();
@@ -2230,11 +2256,12 @@ void main() {
       });
 
       testWidgets('send URLSearchParams data', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer<Map<String, dynamic>> fetchPostCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<void> shouldInterceptFetchPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -2277,9 +2304,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptFetchRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptFetchRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
 
@@ -2311,11 +2338,12 @@ void main() {
       });
 
       testWidgets('send FormData', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer<Map<String, dynamic>> fetchPostCompleter =
-        Completer<Map<String, dynamic>>();
+            Completer<Map<String, dynamic>>();
         final Completer<void> shouldInterceptFetchPostRequestCompleter =
-        Completer<void>();
+            Completer<void>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
@@ -2356,9 +2384,9 @@ void main() {
                     """),
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
-                    clearCache: true,
-                    useShouldInterceptFetchRequest: true,
-                  )),
+                clearCache: true,
+                useShouldInterceptFetchRequest: true,
+              )),
               onWebViewCreated: (controller) {
                 controllerCompleter.complete(controller);
 
@@ -2376,7 +2404,9 @@ void main() {
                 var bodyString = String.fromCharCodes(body);
                 assert(bodyString.indexOf("WebKitFormBoundary") >= 0);
 
-                fetchRequest.body = utf8.encode(bodyString.replaceFirst("Foo", "Foo2").replaceFirst("Bar", "Bar2"));
+                fetchRequest.body = utf8.encode(bodyString
+                    .replaceFirst("Foo", "Foo2")
+                    .replaceFirst("Bar", "Bar2"));
                 shouldInterceptFetchPostRequestCompleter.complete();
                 return fetchRequest;
               },
@@ -2403,24 +2433,24 @@ void main() {
           child: InAppWebView(
             key: GlobalKey(),
             initialUrlRequest:
-            URLRequest(url: Uri.parse('https://flutter.dev/')),
+                URLRequest(url: Uri.parse('https://flutter.dev/')),
             onWebViewCreated: (controller) {
               controllerCompleter.complete(controller);
             },
             initialOptions: InAppWebViewGroupOptions(
                 crossPlatform:
-                InAppWebViewOptions(clearCache: true, contentBlockers: [
-                  ContentBlocker(
-                      trigger:
+                    InAppWebViewOptions(clearCache: true, contentBlockers: [
+              ContentBlocker(
+                  trigger:
                       ContentBlockerTrigger(urlFilter: ".*", resourceType: [
-                        ContentBlockerTriggerResourceType.IMAGE,
-                        ContentBlockerTriggerResourceType.STYLE_SHEET
-                      ], ifTopUrl: [
-                        "https://flutter.dev/"
-                      ]),
-                      action: ContentBlockerAction(
-                          type: ContentBlockerActionType.BLOCK))
-                ])),
+                    ContentBlockerTriggerResourceType.IMAGE,
+                    ContentBlockerTriggerResourceType.STYLE_SHEET
+                  ], ifTopUrl: [
+                    "https://flutter.dev/"
+                  ]),
+                  action: ContentBlockerAction(
+                      type: ContentBlockerActionType.BLOCK))
+            ])),
             onLoadStop: (controller, url) {
               pageLoaded.complete();
             },
@@ -2705,7 +2735,7 @@ void main() {
       expect(numberOfMatches, 2);
     });
 
-    testWidgets('onDownloadStart', (WidgetTester tester) async {
+    testWidgets('onDownloadStartRequest', (WidgetTester tester) async {
       final Completer controllerCompleter = Completer<InAppWebViewController>();
       final Completer<String> onDownloadStartCompleter = Completer<String>();
       await tester.pumpWidget(
@@ -2739,8 +2769,8 @@ void main() {
             onWebViewCreated: (controller) {
               controllerCompleter.complete(controller);
             },
-            onDownloadStart: (controller, url) {
-              onDownloadStartCompleter.complete(url.toString());
+            onDownloadStartRequest: (controller, request) {
+              onDownloadStartCompleter.complete(request.url.toString());
             },
           ),
         ),
@@ -3260,7 +3290,7 @@ setTimeout(function() {
           child: InAppWebView(
             key: GlobalKey(),
             initialUrlRequest:
-            URLRequest(url: Uri.parse('https://github.com/flutter')),
+                URLRequest(url: Uri.parse('https://github.com/flutter')),
             onWebViewCreated: (controller) {
               controllerCompleter.complete(controller);
             },
@@ -3277,7 +3307,7 @@ setTimeout(function() {
       );
 
       final InAppWebViewController controller =
-      await controllerCompleter.future;
+          await controllerCompleter.future;
       await pageLoaded.future;
       listenForScaleChange = true;
 
@@ -3804,7 +3834,7 @@ setTimeout(function() {
 """,
           encoding: 'utf-8',
           mimeType: 'text/html',
-          androidHistoryUrl: Uri.parse("https://flutter.dev"),
+          historyUrl: Uri.parse("https://flutter.dev"),
           baseUrl: Uri.parse("https://flutter.dev"));
       await pageLoads.stream.first;
 
@@ -4133,7 +4163,7 @@ setTimeout(function() {
       if (Platform.isAndroid) {
         await pageLoaded.future;
         expect(await controller.evaluateJavascript(source: "document.body"),
-            isNull);
+            isNullOrEmpty);
       } else if (Platform.isIOS) {
         expect(pageLoaded.future, doesNotComplete);
       }
@@ -4166,10 +4196,14 @@ setTimeout(function() {
       await pageLoaded.future;
 
       await controller.injectJavascriptFileFromUrl(
-          urlFile: Uri.parse('https://www.notawebsite..com/jquery-3.3.1.min.js'),
-          scriptHtmlTagAttributes: ScriptHtmlTagAttributes(id: 'jquery-error', onError: () {
-            jQueryLoadError.complete();
-          },));
+          urlFile:
+              Uri.parse('https://www.notawebsite..com/jquery-3.3.1.min.js'),
+          scriptHtmlTagAttributes: ScriptHtmlTagAttributes(
+            id: 'jquery-error',
+            onError: () {
+              jQueryLoadError.complete();
+            },
+          ));
       await jQueryLoadError.future;
       expect(
           await controller.evaluateJavascript(
@@ -4181,9 +4215,12 @@ setTimeout(function() {
 
       await controller.injectJavascriptFileFromUrl(
           urlFile: Uri.parse('https://code.jquery.com/jquery-3.3.1.min.js'),
-          scriptHtmlTagAttributes: ScriptHtmlTagAttributes(id: 'jquery', onLoad: () {
-            jQueryLoaded.complete();
-          },));
+          scriptHtmlTagAttributes: ScriptHtmlTagAttributes(
+            id: 'jquery',
+            onLoad: () {
+              jQueryLoaded.complete();
+            },
+          ));
       await jQueryLoaded.future;
       expect(
           await controller.evaluateJavascript(
@@ -4550,7 +4587,7 @@ setTimeout(function() {
       await pageLoaded.future;
 
       await expectLater(
-          controller.zoomBy(zoomFactor: 3.0, iosAnimated: true), completes);
+          controller.zoomBy(zoomFactor: 3.0, animated: true), completes);
     });
 
     testWidgets('getZoomScale', (WidgetTester tester) async {
@@ -4853,7 +4890,7 @@ setTimeout(function() {
         fileName = fileName + WebArchiveFormat.WEBARCHIVE.toValue();
       }
 
-      var fullPath = supportDir!.path + Platform.pathSeparator + fileName;
+      var fullPath = supportDir.path + Platform.pathSeparator + fileName;
       var path = await controller.saveWebArchive(filePath: fullPath);
       expect(path, isNotNull);
       expect(path, endsWith(fileName));
@@ -4902,21 +4939,19 @@ setTimeout(function() {
       expect(await InAppWebViewController.getDefaultUserAgent(), isNotNull);
     });
 
-    testWidgets('launches with pull-to-refresh feature', (WidgetTester tester) async {
+    testWidgets('launches with pull-to-refresh feature',
+        (WidgetTester tester) async {
       final Completer controllerCompleter = Completer<InAppWebViewController>();
       final pullToRefreshController = PullToRefreshController(
         options: PullToRefreshOptions(
-          color: Colors.blue,
-          size: AndroidPullToRefreshSize.DEFAULT,
-          backgroundColor: Colors.grey,
-          enabled: true,
-          slingshotDistance: 150,
-          distanceToTriggerSync: 150,
-          attributedTitle: IOSNSAttributedString(string: "test")
-        ),
-        onRefresh: () {
-
-        },
+            color: Colors.blue,
+            size: AndroidPullToRefreshSize.DEFAULT,
+            backgroundColor: Colors.grey,
+            enabled: true,
+            slingshotDistance: 150,
+            distanceToTriggerSync: 150,
+            attributedTitle: IOSNSAttributedString(string: "test")),
+        onRefresh: () {},
       );
 
       await tester.pumpWidget(
@@ -4924,12 +4959,11 @@ setTimeout(function() {
           textDirection: TextDirection.ltr,
           child: InAppWebView(
             key: GlobalKey(),
-            initialUrlRequest: URLRequest(url: Uri.parse('https://github.com/flutter')),
+            initialUrlRequest:
+                URLRequest(url: Uri.parse('https://github.com/flutter')),
             initialOptions: InAppWebViewGroupOptions(
-              android: AndroidInAppWebViewOptions(
-                useHybridComposition: true
-              )
-            ),
+                android:
+                    AndroidInAppWebViewOptions(useHybridComposition: true)),
             pullToRefreshController: pullToRefreshController,
             onWebViewCreated: (controller) {
               controllerCompleter.complete(controller);
@@ -4938,22 +4972,22 @@ setTimeout(function() {
         ),
       );
       final InAppWebViewController controller =
-      await controllerCompleter.future;
+          await controllerCompleter.future;
       final String? currentUrl = (await controller.getUrl())?.toString();
       expect(currentUrl, 'https://github.com/flutter');
     });
 
     group('WebMessage', () {
       testWidgets('WebMessageChannel', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer webMessageCompleter = Completer<String>();
         await tester.pumpWidget(
           Directionality(
             textDirection: TextDirection.ltr,
             child: InAppWebView(
               key: GlobalKey(),
-              initialData: InAppWebViewInitialData(
-                  data: """
+              initialData: InAppWebViewInitialData(data: """
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -4988,15 +5022,20 @@ setTimeout(function() {
                 webMessageCompleter.complete(consoleMessage.message);
               },
               onLoadStop: (controller, url) async {
-                var webMessageChannel = await controller.createWebMessageChannel();
+                var webMessageChannel =
+                    await controller.createWebMessageChannel();
                 var port1 = webMessageChannel!.port1;
                 var port2 = webMessageChannel.port2;
 
                 await port1.setWebMessageCallback((message) async {
-                  await port1.postMessage(WebMessage(data: message! + " and back"));
+                  await port1
+                      .postMessage(WebMessage(data: message! + " and back"));
                 });
-                await controller.postWebMessage(message: WebMessage(data: "capturePort", ports: [port2]), targetOrigin: Uri.parse("*"));
-                await controller.evaluateJavascript(source: "document.getElementById('button').click();");
+                await controller.postWebMessage(
+                    message: WebMessage(data: "capturePort", ports: [port2]),
+                    targetOrigin: Uri.parse("*"));
+                await controller.evaluateJavascript(
+                    source: "document.getElementById('button').click();");
               },
             ),
           ),
@@ -5008,7 +5047,8 @@ setTimeout(function() {
       });
 
       testWidgets('WebMessageListener', (WidgetTester tester) async {
-        final Completer controllerCompleter = Completer<InAppWebViewController>();
+        final Completer controllerCompleter =
+            Completer<InAppWebViewController>();
         final Completer<void> pageLoaded = Completer<void>();
         final Completer webMessageCompleter = Completer<String>();
         await tester.pumpWidget(
@@ -5020,8 +5060,10 @@ setTimeout(function() {
                 await controller.addWebMessageListener(WebMessageListener(
                   jsObjectName: "myTestObj",
                   allowedOriginRules: Set.from(["https://*.example.com"]),
-                  onPostMessage: (message, sourceOrigin, isMainFrame, replyProxy) {
-                    assert(sourceOrigin.toString() == "https://www.example.com");
+                  onPostMessage:
+                      (message, sourceOrigin, isMainFrame, replyProxy) {
+                    assert(
+                        sourceOrigin.toString() == "https://www.example.com");
                     assert(isMainFrame);
 
                     replyProxy.postMessage(message! + " and back");
@@ -5041,7 +5083,8 @@ setTimeout(function() {
           ),
         );
         final controller = await controllerCompleter.future;
-        await controller.loadUrl(urlRequest: URLRequest(url: Uri.parse("https://www.example.com/")));
+        await controller.loadUrl(
+            urlRequest: URLRequest(url: Uri.parse("https://www.example.com/")));
         await pageLoaded.future;
 
         await controller.evaluateJavascript(source: """
@@ -5140,9 +5183,9 @@ setTimeout(function() {
         final InAppWebViewController controller =
             await controllerCompleter.future;
         await pageLoaded.future;
-        var originUrl = (await controller.android.getOriginalUrl())?.toString();
+        var originUrl = (await controller.getOriginalUrl())?.toString();
         expect(originUrl, 'https://github.com/flutter');
-      }, skip: !Platform.isAndroid);
+      });
 
       testWidgets('pageDown/pageUp', (WidgetTester tester) async {
         final Completer controllerCompleter =
@@ -5270,6 +5313,12 @@ setTimeout(function() {
         expect(await AndroidInAppWebViewController.getCurrentWebViewPackage(),
             isNotNull);
       }, skip: !Platform.isAndroid);
+
+      test('setWebContentsDebuggingEnabled', () async {
+        expect(
+            AndroidInAppWebViewController.setWebContentsDebuggingEnabled(true),
+            completes);
+      }, skip: !Platform.isAndroid);
     }, skip: !Platform.isAndroid);
 
     group('ios methods', () {
@@ -5372,8 +5421,7 @@ setTimeout(function() {
             textDirection: TextDirection.ltr,
             child: InAppWebView(
               key: GlobalKey(),
-              initialData: InAppWebViewInitialData(
-                  data: """
+              initialData: InAppWebViewInitialData(data: """
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -5388,8 +5436,7 @@ setTimeout(function() {
     </script>
 </body>
 </html>
-                  """
-              ),
+                  """),
               initialOptions: InAppWebViewGroupOptions(
                 ios: IOSInAppWebViewOptions(
                   applePayAPIEnabled: true,
@@ -5414,6 +5461,80 @@ setTimeout(function() {
         expect(await IOSInAppWebViewController.handlesURLScheme("https"), true);
       }, skip: !Platform.isIOS);
     }, skip: !Platform.isIOS);
+  });
+
+  group('Service Worker', () {
+    testWidgets('shouldInterceptRequest', (WidgetTester tester) async {
+      final Completer completer = Completer();
+
+      var swAvailable = await AndroidWebViewFeature.isFeatureSupported(
+          AndroidWebViewFeature.SERVICE_WORKER_BASIC_USAGE);
+      var swInterceptAvailable = await AndroidWebViewFeature.isFeatureSupported(
+          AndroidWebViewFeature.SERVICE_WORKER_SHOULD_INTERCEPT_REQUEST);
+
+      if (swAvailable && swInterceptAvailable) {
+        AndroidServiceWorkerController serviceWorkerController =
+            AndroidServiceWorkerController.instance();
+
+        await serviceWorkerController
+            .setServiceWorkerClient(AndroidServiceWorkerClient(
+          shouldInterceptRequest: (request) async {
+            if (!completer.isCompleted) {
+              completer.complete();
+            }
+            return null;
+          },
+        ));
+      } else {
+        completer.complete();
+      }
+
+      await tester.pumpWidget(
+        Directionality(
+          textDirection: TextDirection.ltr,
+          child: InAppWebView(
+            key: GlobalKey(),
+            initialUrlRequest:
+                URLRequest(url: Uri.parse('https://mdn.github.io/sw-test/')),
+          ),
+        ),
+      );
+
+      expect(completer.future, completes);
+    }, skip: !Platform.isAndroid);
+
+    testWidgets('setServiceWorkerClient to null', (WidgetTester tester) async {
+      final Completer<String> pageLoaded = Completer<String>();
+
+      var swAvailable = await AndroidWebViewFeature.isFeatureSupported(
+          AndroidWebViewFeature.SERVICE_WORKER_BASIC_USAGE);
+      var swInterceptAvailable = await AndroidWebViewFeature.isFeatureSupported(
+          AndroidWebViewFeature.SERVICE_WORKER_SHOULD_INTERCEPT_REQUEST);
+
+      if (swAvailable && swInterceptAvailable) {
+        AndroidServiceWorkerController serviceWorkerController =
+            AndroidServiceWorkerController.instance();
+
+        await serviceWorkerController.setServiceWorkerClient(null);
+      }
+
+      await tester.pumpWidget(
+        Directionality(
+          textDirection: TextDirection.ltr,
+          child: InAppWebView(
+            key: GlobalKey(),
+            initialUrlRequest:
+                URLRequest(url: Uri.parse('https://mdn.github.io/sw-test/')),
+            onLoadStop: (controller, url) {
+              pageLoaded.complete(url!.toString());
+            },
+          ),
+        ),
+      );
+
+      final String url = await pageLoaded.future;
+      expect(url, "https://mdn.github.io/sw-test/");
+    }, skip: !Platform.isAndroid);
   });
 
   group('Cookie Manager', () {
@@ -5469,7 +5590,8 @@ setTimeout(function() {
       final Completer<void> pageLoaded = Completer<void>();
 
       var headlessWebView = new HeadlessInAppWebView(
-        initialUrlRequest: URLRequest(url: Uri.parse("https://github.com/flutter")),
+        initialUrlRequest:
+            URLRequest(url: Uri.parse("https://github.com/flutter")),
         onWebViewCreated: (controller) {
           controllerCompleter.complete(controller);
         },
@@ -5498,20 +5620,20 @@ setTimeout(function() {
       final Completer<void> pageLoaded = Completer<void>();
 
       var headlessWebView = new HeadlessInAppWebView(
-        initialUrlRequest: URLRequest(url: Uri.parse("https://github.com/flutter")),
-        onWebViewCreated: (controller) {
-          controllerCompleter.complete(controller);
-        },
-        onLoadStop: (controller, url) async {
-          pageLoaded.complete();
-        }
-      );
+          initialUrlRequest:
+              URLRequest(url: Uri.parse("https://github.com/flutter")),
+          onWebViewCreated: (controller) {
+            controllerCompleter.complete(controller);
+          },
+          onLoadStop: (controller, url) async {
+            pageLoaded.complete();
+          });
 
       await headlessWebView.run();
       expect(headlessWebView.isRunning(), true);
 
       final InAppWebViewController controller =
-      await controllerCompleter.future;
+          await controllerCompleter.future;
       await pageLoaded.future;
 
       final String? url = (await controller.getUrl())?.toString();
@@ -5532,11 +5654,12 @@ setTimeout(function() {
       final Completer controllerCompleter = Completer<InAppWebViewController>();
 
       var headlessWebView = new HeadlessInAppWebView(
-          initialUrlRequest: URLRequest(url: Uri.parse("https://github.com/flutter")),
-          initialSize: Size(600, 800),
-          onWebViewCreated: (controller) {
-            controllerCompleter.complete(controller);
-          },
+        initialUrlRequest:
+            URLRequest(url: Uri.parse("https://github.com/flutter")),
+        initialSize: Size(600, 800),
+        onWebViewCreated: (controller) {
+          controllerCompleter.complete(controller);
+        },
       );
 
       await headlessWebView.run();
@@ -5561,7 +5684,8 @@ setTimeout(function() {
       final Completer<void> pageLoaded = Completer<void>();
 
       var headlessWebView = new HeadlessInAppWebView(
-        initialUrlRequest: URLRequest(url: Uri.parse("https://github.com/flutter")),
+        initialUrlRequest:
+            URLRequest(url: Uri.parse("https://github.com/flutter")),
         initialOptions: InAppWebViewGroupOptions(
             crossPlatform: InAppWebViewOptions(javaScriptEnabled: false)),
         onWebViewCreated: (controller) {
@@ -5605,8 +5729,7 @@ setTimeout(function() {
       expect(inAppBrowser.isOpened(), true);
       expect(() async {
         await inAppBrowser.openUrlRequest(
-            urlRequest:
-                URLRequest(url: Uri.parse("https://flutter.dev")));
+            urlRequest: URLRequest(url: Uri.parse("https://flutter.dev")));
       }, throwsA(isInstanceOf<InAppBrowserAlreadyOpenedException>()));
 
       await inAppBrowser.firstPageLoaded.future;
@@ -5628,13 +5751,14 @@ setTimeout(function() {
         await inAppBrowser.show();
       }, throwsA(isInstanceOf<InAppBrowserNotOpenedException>()));
 
-      await inAppBrowser.openFile(assetFilePath: "test_assets/in_app_webview_initial_file_test.html");
+      await inAppBrowser.openFile(
+          assetFilePath: "test_assets/in_app_webview_initial_file_test.html");
       await inAppBrowser.browserCreated.future;
       expect(inAppBrowser.isOpened(), true);
       expect(() async {
         await inAppBrowser.openUrlRequest(
             urlRequest:
-            URLRequest(url: Uri.parse("https://github.com/flutter")));
+                URLRequest(url: Uri.parse("https://github.com/flutter")));
       }, throwsA(isInstanceOf<InAppBrowserAlreadyOpenedException>()));
 
       await inAppBrowser.firstPageLoaded.future;
@@ -5656,7 +5780,8 @@ setTimeout(function() {
         await inAppBrowser.show();
       }, throwsA(isInstanceOf<InAppBrowserNotOpenedException>()));
 
-      await inAppBrowser.openData(data: """
+      await inAppBrowser.openData(
+          data: """
 <!doctype html>
 <html lang="en">
     <head>
@@ -5680,7 +5805,7 @@ setTimeout(function() {
       expect(() async {
         await inAppBrowser.openUrlRequest(
             urlRequest:
-            URLRequest(url: Uri.parse("https://github.com/flutter")));
+                URLRequest(url: Uri.parse("https://github.com/flutter")));
       }, throwsA(isInstanceOf<InAppBrowserAlreadyOpenedException>()));
 
       await inAppBrowser.firstPageLoaded.future;
@@ -5723,12 +5848,12 @@ setTimeout(function() {
       var chromeSafariBrowser = new MyChromeSafariBrowser();
       expect(chromeSafariBrowser.isOpened(), false);
 
-      await chromeSafariBrowser.open(url: Uri.parse("https://github.com/flutter"));
+      await chromeSafariBrowser.open(
+          url: Uri.parse("https://github.com/flutter"));
       await chromeSafariBrowser.browserCreated.future;
       expect(chromeSafariBrowser.isOpened(), true);
       expect(() async {
-        await chromeSafariBrowser.open(
-            url: Uri.parse("https://flutter.dev"));
+        await chromeSafariBrowser.open(url: Uri.parse("https://flutter.dev"));
       }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
 
       await expectLater(chromeSafariBrowser.firstPageLoaded.future, completes);
@@ -5736,6 +5861,126 @@ setTimeout(function() {
       await chromeSafariBrowser.browserClosed.future;
       expect(chromeSafariBrowser.isOpened(), false);
     });
+
+    test('add custom menu item', () async {
+      var chromeSafariBrowser = new MyChromeSafariBrowser();
+      chromeSafariBrowser.addMenuItem(ChromeSafariBrowserMenuItem(
+          id: 2,
+          label: 'Custom item menu 1',
+          action: (url, title) {
+            print('Custom item menu 1 clicked!');
+          }));
+      expect(chromeSafariBrowser.isOpened(), false);
+
+      await chromeSafariBrowser.open(
+          url: Uri.parse("https://github.com/flutter"));
+      await chromeSafariBrowser.browserCreated.future;
+      expect(chromeSafariBrowser.isOpened(), true);
+      expect(() async {
+        await chromeSafariBrowser.open(url: Uri.parse("https://flutter.dev"));
+      }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
+
+      await expectLater(chromeSafariBrowser.firstPageLoaded.future, completes);
+      await chromeSafariBrowser.close();
+      await chromeSafariBrowser.browserClosed.future;
+      expect(chromeSafariBrowser.isOpened(), false);
+    });
+
+    group('Android Custom Tabs', () {
+      test('add custom action button', () async {
+        var chromeSafariBrowser = new MyChromeSafariBrowser();
+        var actionButtonIcon =
+            await rootBundle.load('test_assets/images/flutter-logo.png');
+        chromeSafariBrowser.setActionButton(ChromeSafariBrowserActionButton(
+            id: 1,
+            description: 'Action Button description',
+            icon: actionButtonIcon.buffer.asUint8List(),
+            action: (url, title) {
+              print('Action Button 1 clicked!');
+            }));
+        expect(chromeSafariBrowser.isOpened(), false);
+
+        await chromeSafariBrowser.open(
+            url: Uri.parse("https://github.com/flutter"));
+        await chromeSafariBrowser.browserCreated.future;
+        expect(chromeSafariBrowser.isOpened(), true);
+        expect(() async {
+          await chromeSafariBrowser.open(url: Uri.parse("https://flutter.dev"));
+        }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
+
+        await expectLater(
+            chromeSafariBrowser.firstPageLoaded.future, completes);
+        await chromeSafariBrowser.close();
+        await chromeSafariBrowser.browserClosed.future;
+        expect(chromeSafariBrowser.isOpened(), false);
+      });
+
+      test('Custom Tabs single instance', () async {
+        var chromeSafariBrowser = new MyChromeSafariBrowser();
+        expect(chromeSafariBrowser.isOpened(), false);
+
+        await chromeSafariBrowser.open(
+            url: Uri.parse("https://github.com/flutter"),
+            options: ChromeSafariBrowserClassOptions(
+                android:
+                    AndroidChromeCustomTabsOptions(isSingleInstance: true)));
+        await chromeSafariBrowser.browserCreated.future;
+        expect(chromeSafariBrowser.isOpened(), true);
+        expect(() async {
+          await chromeSafariBrowser.open(url: Uri.parse("https://flutter.dev"));
+        }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
+
+        await expectLater(
+            chromeSafariBrowser.firstPageLoaded.future, completes);
+        await chromeSafariBrowser.close();
+        await chromeSafariBrowser.browserClosed.future;
+        expect(chromeSafariBrowser.isOpened(), false);
+      });
+
+      test('Trusted Web Activity', () async {
+        var chromeSafariBrowser = new MyChromeSafariBrowser();
+        expect(chromeSafariBrowser.isOpened(), false);
+
+        await chromeSafariBrowser.open(
+            url: Uri.parse("https://github.com/flutter"),
+            options: ChromeSafariBrowserClassOptions(
+                android: AndroidChromeCustomTabsOptions(
+                    isTrustedWebActivity: true)));
+        await chromeSafariBrowser.browserCreated.future;
+        expect(chromeSafariBrowser.isOpened(), true);
+        expect(() async {
+          await chromeSafariBrowser.open(url: Uri.parse("https://flutter.dev"));
+        }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
+
+        await expectLater(
+            chromeSafariBrowser.firstPageLoaded.future, completes);
+        await chromeSafariBrowser.close();
+        await chromeSafariBrowser.browserClosed.future;
+        expect(chromeSafariBrowser.isOpened(), false);
+      });
+
+      test('Trusted Web Activity single instance', () async {
+        var chromeSafariBrowser = new MyChromeSafariBrowser();
+        expect(chromeSafariBrowser.isOpened(), false);
+
+        await chromeSafariBrowser.open(
+            url: Uri.parse("https://github.com/flutter"),
+            options: ChromeSafariBrowserClassOptions(
+                android: AndroidChromeCustomTabsOptions(
+                    isTrustedWebActivity: true, isSingleInstance: true)));
+        await chromeSafariBrowser.browserCreated.future;
+        expect(chromeSafariBrowser.isOpened(), true);
+        expect(() async {
+          await chromeSafariBrowser.open(url: Uri.parse("https://flutter.dev"));
+        }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
+
+        await expectLater(
+            chromeSafariBrowser.firstPageLoaded.future, completes);
+        await chromeSafariBrowser.close();
+        await chromeSafariBrowser.browserClosed.future;
+        expect(chromeSafariBrowser.isOpened(), false);
+      });
+    }, skip: !Platform.isAndroid);
   });
 
   group('InAppLocalhostServer', () {
@@ -5754,8 +5999,8 @@ setTimeout(function() {
           textDirection: TextDirection.ltr,
           child: InAppWebView(
             key: GlobalKey(),
-            initialUrlRequest:
-            URLRequest(url: Uri.parse('http://localhost:8080/test_assets/index.html')),
+            initialUrlRequest: URLRequest(
+                url: Uri.parse('http://localhost:8080/test_assets/index.html')),
             onWebViewCreated: (controller) {
               controllerCompleter.complete(controller);
             },
@@ -5763,7 +6008,7 @@ setTimeout(function() {
         ),
       );
       final InAppWebViewController controller =
-      await controllerCompleter.future;
+          await controllerCompleter.future;
       final String? currentUrl = (await controller.getUrl())?.toString();
       expect(currentUrl, 'http://localhost:8080/test_assets/index.html');
     });
