@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 
 import 'main.dart';
@@ -32,8 +33,20 @@ class _ChromeSafariBrowserExampleScreenState
     extends State<ChromeSafariBrowserExampleScreen> {
   @override
   void initState() {
+    rootBundle.load('assets/images/flutter-logo.png').then((actionButtonIcon) {
+      widget.browser.setActionButton(ChromeSafariBrowserActionButton(
+          id: 1,
+          description: 'Action Button description',
+          icon: actionButtonIcon.buffer.asUint8List(),
+          action: (url, title) {
+            print('Action Button 1 clicked!');
+            print(url);
+            print(title);
+          }));
+    });
+
     widget.browser.addMenuItem(ChromeSafariBrowserMenuItem(
-        id: 1,
+        id: 2,
         label: 'Custom item menu 1',
         action: (url, title) {
           print('Custom item menu 1 clicked!');
@@ -41,7 +54,7 @@ class _ChromeSafariBrowserExampleScreenState
           print(title);
         }));
     widget.browser.addMenuItem(ChromeSafariBrowserMenuItem(
-        id: 2,
+        id: 3,
         label: 'Custom item menu 2',
         action: (url, title) {
           print('Custom item menu 2 clicked!');
@@ -66,7 +79,9 @@ class _ChromeSafariBrowserExampleScreenState
                     url: Uri.parse("https://flutter.dev/"),
                     options: ChromeSafariBrowserClassOptions(
                         android: AndroidChromeCustomTabsOptions(
-                            addDefaultShareMenuItem: false,
+                            shareState: CustomTabsShareState.SHARE_STATE_OFF,
+                            isSingleInstance: false,
+                            isTrustedWebActivity: false,
                             keepAliveEnabled: true),
                         ios: IOSSafariOptions(
                             dismissButtonStyle:
