@@ -1,6 +1,8 @@
 import 'dart:async';
 import 'dart:collection';
+import 'dart:typed_data';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_inappwebview/src/util.dart';
 
@@ -40,6 +42,7 @@ class ChromeSafariBrowser {
   ///View ID used internally.
   late final String id;
 
+  ChromeSafariBrowserActionButton? _actionButton;
   Map<int, ChromeSafariBrowserMenuItem> _menuItems = new HashMap();
   bool _isOpened = false;
   late MethodChannel _channel;
@@ -66,12 +69,14 @@ class ChromeSafariBrowser {
         onClosed();
         this._isOpened = false;
         break;
-      case "onChromeSafariBrowserMenuItemActionPerform":
+      case "onChromeSafariBrowserItemActionPerform":
         String url = call.arguments["url"];
         String title = call.arguments["title"];
         int id = call.arguments["id"].toInt();
-        if (this._menuItems[id] != null) {
-          this._menuItems[id]!.action(url, title);
+        if (this._actionButton?.id == id) {
+          this._actionButton?.action(url, title);
+        } else if (this._menuItems[id] != null) {
+          this._menuItems[id]?.action(url, title);
         }
         break;
       default:
@@ -81,23 +86,28 @@ class ChromeSafariBrowser {
 
   ///Opens the [ChromeSafariBrowser] instance with an [url].
   ///
-  ///[url]: The [url] to load.
+  ///[url]: The [url] to load. On iOS, the [url] must use the `http` or `https` scheme.
   ///
   ///[options]: Options for the [ChromeSafariBrowser].
   Future<void> open(
       {required Uri url, ChromeSafariBrowserClassOptions? options}) async {
     assert(url.toString().isNotEmpty);
     this.throwIsAlreadyOpened(message: 'Cannot open $url!');
+    if (!kIsWeb && defaultTargetPlatform == TargetPlatform.iOS) {
+      assert(['http', 'https'].contains(url.scheme),
+          'The specified URL has an unsupported scheme. Only HTTP and HTTPS URLs are supported on iOS.');
+    }
 
     List<Map<String, dynamic>> menuItemList = [];
     _menuItems.forEach((key, value) {
-      menuItemList.add({"id": value.id, "label": value.label});
+      menuItemList.add(value.toMap());
     });
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('id', () => id);
     args.putIfAbsent('url', () => url.toString());
     args.putIfAbsent('options', () => options?.toMap() ?? {});
+    args.putIfAbsent('actionButton', () => _actionButton?.toMap());
     args.putIfAbsent('menuItemList', () => menuItemList);
     await _sharedChannel.invokeMethod('open', args);
     this._isOpened = true;
@@ -109,12 +119,26 @@ class ChromeSafariBrowser {
     await _channel.invokeMethod("close", args);
   }
 
+  ///Set a custom action button.
+  ///
+  ///**NOTE**: Not available in a Trusted Web Activity.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android ([Official API - CustomTabsIntent.Builder.setActionButton ](https://developer.android.com/reference/androidx/browser/customtabs/CustomTabsIntent.Builder#setActionButton(android.graphics.Bitmap,%20java.lang.String,%20android.app.PendingIntent,%20boolean)))
+  void setActionButton(ChromeSafariBrowserActionButton actionButton) {
+    this._actionButton = actionButton;
+  }
+
   ///Adds a [ChromeSafariBrowserMenuItem] to the menu.
+  ///
+  ///**NOTE**: Not available in an Android Trusted Web Activity.
   void addMenuItem(ChromeSafariBrowserMenuItem menuItem) {
     this._menuItems[menuItem.id] = menuItem;
   }
 
   ///Adds a list of [ChromeSafariBrowserMenuItem] to the menu.
+  ///
+  ///**NOTE**: Not available in an Android Trusted Web Activity.
   void addMenuItems(List<ChromeSafariBrowserMenuItem> menuItems) {
     menuItems.forEach((menuItem) {
       this._menuItems[menuItem.id] = menuItem;
@@ -160,9 +184,56 @@ class ChromeSafariBrowser {
   }
 }
 
+///Class that represents a custom action button for a [ChromeSafariBrowser] instance.
+///
+///**NOTE**: Not available in an Android Trusted Web Activity.
+class ChromeSafariBrowserActionButton {
+  ///The action button id. It should be different from the [ChromeSafariBrowserMenuItem.id].
+  int id;
+
+  ///The icon byte data.
+  Uint8List icon;
+
+  ///The description for the button. To be used for accessibility.
+  String description;
+
+  ///Whether the action button should be tinted.
+  bool shouldTint;
+
+  ///Callback function to be invoked when the menu item is clicked
+  final void Function(String url, String title) action;
+
+  ChromeSafariBrowserActionButton(
+      {required this.id,
+      required this.icon,
+      required this.description,
+      required this.action,
+      this.shouldTint = false});
+
+  Map<String, dynamic> toMap() {
+    return {
+      "id": id,
+      "icon": icon,
+      "description": description,
+      "shouldTint": shouldTint
+    };
+  }
+
+  Map<String, dynamic> toJson() {
+    return this.toMap();
+  }
+
+  @override
+  String toString() {
+    return toMap().toString();
+  }
+}
+
 ///Class that represents a custom menu item for a [ChromeSafariBrowser] instance.
+///
+///**NOTE**: Not available in an Android Trusted Web Activity.
 class ChromeSafariBrowserMenuItem {
-  ///The menu item id
+  ///The menu item id. It should be different from [ChromeSafariBrowserActionButton.id].
   int id;
 
   ///The label of the menu item
