@@ -8,7 +8,6 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter/gestures.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 
 import '../context_menu.dart';
 import '../types.dart';
@@ -17,6 +16,7 @@ import 'webview.dart';
 import 'in_app_webview_controller.dart';
 import 'in_app_webview_options.dart';
 import '../pull_to_refresh/pull_to_refresh_controller.dart';
+import '../pull_to_refresh/pull_to_refresh_options.dart';
 
 ///Flutter Widget for adding an **inline native WebView** integrated in the flutter widget tree.
 class InAppWebView extends StatefulWidget implements WebView {
@@ -41,6 +41,7 @@ class InAppWebView extends StatefulWidget implements WebView {
     this.initialOptions,
     this.initialUserScripts,
     this.pullToRefreshController,
+    this.implementation = WebViewImplementation.NATIVE,
     this.contextMenu,
     this.onWebViewCreated,
     this.onLoadStart,
@@ -52,7 +53,8 @@ class InAppWebView extends StatefulWidget implements WebView {
     this.shouldOverrideUrlLoading,
     this.onLoadResource,
     this.onScrollChanged,
-    this.onDownloadStart,
+    @Deprecated('Use `onDownloadStartRequest` instead') this.onDownloadStart,
+    this.onDownloadStartRequest,
     this.onLoadResourceCustomScheme,
     this.onCreateWindow,
     this.onCloseWindow,
@@ -136,6 +138,9 @@ class InAppWebView extends StatefulWidget implements WebView {
   final URLRequest? initialUrlRequest;
 
   @override
+  final WebViewImplementation implementation;
+
+  @override
   final UnmodifiableListView<UserScript>? initialUserScripts;
 
   @override
@@ -207,9 +212,15 @@ class InAppWebView extends StatefulWidget implements WebView {
           InAppWebViewController controller, Uri url, bool precomposed)?
       androidOnReceivedTouchIconUrl;
 
+  ///Use [onDownloadStartRequest] instead
+  @Deprecated('Use `onDownloadStartRequest` instead')
   @override
   final void Function(InAppWebViewController controller, Uri url)?
       onDownloadStart;
+
+  @override
+  final void Function(InAppWebViewController controller,
+      DownloadStartRequest downloadStartRequest)? onDownloadStartRequest;
 
   @override
   final void Function(InAppWebViewController controller, int activeMatchOrdinal,
@@ -395,14 +406,13 @@ class _InAppWebViewState extends State<InAppWebView> {
               viewType: 'com.pichillilorenzo/flutter_inappwebview',
               layoutDirection: TextDirection.rtl,
               creationParams: <String, dynamic>{
-                'initialUrlRequest': (widget.initialUrlRequest ??
-                        URLRequest(url: Uri.parse("about:blank")))
-                    .toMap(),
+                'initialUrlRequest': widget.initialUrlRequest?.toMap(),
                 'initialFile': widget.initialFile,
                 'initialData': widget.initialData?.toMap(),
                 'initialOptions': widget.initialOptions?.toMap() ?? {},
                 'contextMenu': widget.contextMenu?.toMap() ?? {},
                 'windowId': widget.windowId,
+                'implementation': widget.implementation.toValue(),
                 'initialUserScripts':
                     widget.initialUserScripts?.map((e) => e.toMap()).toList() ??
                         [],
@@ -423,16 +433,15 @@ class _InAppWebViewState extends State<InAppWebView> {
           viewType: 'com.pichillilorenzo/flutter_inappwebview',
           onPlatformViewCreated: _onPlatformViewCreated,
           gestureRecognizers: widget.gestureRecognizers,
-          layoutDirection: TextDirection.rtl,
+          layoutDirection: Directionality.maybeOf(context) ?? TextDirection.rtl,
           creationParams: <String, dynamic>{
-            'initialUrlRequest': (widget.initialUrlRequest ??
-                    URLRequest(url: Uri.parse("about:blank")))
-                .toMap(),
+            'initialUrlRequest': widget.initialUrlRequest?.toMap(),
             'initialFile': widget.initialFile,
             'initialData': widget.initialData?.toMap(),
             'initialOptions': widget.initialOptions?.toMap() ?? {},
             'contextMenu': widget.contextMenu?.toMap() ?? {},
             'windowId': widget.windowId,
+            'implementation': widget.implementation.toValue(),
             'initialUserScripts':
                 widget.initialUserScripts?.map((e) => e.toMap()).toList() ?? [],
             'pullToRefreshOptions':
@@ -448,14 +457,13 @@ class _InAppWebViewState extends State<InAppWebView> {
         onPlatformViewCreated: _onPlatformViewCreated,
         gestureRecognizers: widget.gestureRecognizers,
         creationParams: <String, dynamic>{
-          'initialUrlRequest': (widget.initialUrlRequest ??
-                  URLRequest(url: Uri.parse("about:blank")))
-              .toMap(),
+          'initialUrlRequest': widget.initialUrlRequest?.toMap(),
           'initialFile': widget.initialFile,
           'initialData': widget.initialData?.toMap(),
           'initialOptions': widget.initialOptions?.toMap() ?? {},
           'contextMenu': widget.contextMenu?.toMap() ?? {},
           'windowId': widget.windowId,
+          'implementation': widget.implementation.toValue(),
           'initialUserScripts':
               widget.initialUserScripts?.map((e) => e.toMap()).toList() ?? [],
           'pullToRefreshOptions':
