@@ -201,15 +201,29 @@ class InAppWebViewController {
             _inAppBrowser!.onScrollChanged(x, y);
         }
         break;
-      case "onDownloadStart":
-        if ((_webview != null && _webview!.onDownloadStart != null) ||
+      case "onDownloadStartRequest":
+        if ((_webview != null &&
+                // ignore: deprecated_member_use_from_same_package
+                (_webview!.onDownloadStart != null ||
+                    _webview!.onDownloadStartRequest != null)) ||
             _inAppBrowser != null) {
-          String url = call.arguments["url"];
-          Uri uri = Uri.parse(url);
-          if (_webview != null && _webview!.onDownloadStart != null)
-            _webview!.onDownloadStart!(this, uri);
-          else
-            _inAppBrowser!.onDownloadStart(uri);
+          Map<String, dynamic> arguments =
+              call.arguments.cast<String, dynamic>();
+          DownloadStartRequest downloadStartRequest =
+              DownloadStartRequest.fromMap(arguments)!;
+
+          if (_webview != null) {
+            if (_webview!.onDownloadStartRequest != null)
+              _webview!.onDownloadStartRequest!(this, downloadStartRequest);
+            else {
+              // ignore: deprecated_member_use_from_same_package
+              _webview!.onDownloadStart!(this, downloadStartRequest.url);
+            }
+          } else {
+            // ignore: deprecated_member_use_from_same_package
+            _inAppBrowser!.onDownloadStart(downloadStartRequest.url);
+            _inAppBrowser!.onDownloadStartRequest(downloadStartRequest);
+          }
         }
         break;
       case "onLoadResourceCustomScheme":
@@ -375,6 +389,7 @@ class InAppWebViewController {
           } else {
             // ignore: deprecated_member_use_from_same_package
             _inAppBrowser!.androidOnScaleChanged(oldScale, newScale);
+            _inAppBrowser!.onZoomScaleChanged(oldScale, newScale);
           }
         }
         break;
@@ -904,9 +919,9 @@ class InAppWebViewController {
   ///Gets the URL for the current page.
   ///This is not always the same as the URL passed to [WebView.onLoadStart] because although the load for that URL has begun, the current page may not have changed.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#getUrl()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1415005-url
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getUrl](https://developer.android.com/reference/android/webkit/WebView#getUrl()))
+  ///- iOS ([Official API - WKWebView.url](https://developer.apple.com/documentation/webkit/wkwebview/1415005-url))
   Future<Uri?> getUrl() async {
     Map<String, dynamic> args = <String, dynamic>{};
     String? url = await _channel.invokeMethod('getUrl', args);
@@ -915,9 +930,9 @@ class InAppWebViewController {
 
   ///Gets the title for the current page.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#getTitle()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1415015-title
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getTitle](https://developer.android.com/reference/android/webkit/WebView#getTitle()))
+  ///- iOS ([Official API - WKWebView.title](https://developer.apple.com/documentation/webkit/wkwebview/1415015-title))
   Future<String?> getTitle() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('getTitle', args);
@@ -925,9 +940,9 @@ class InAppWebViewController {
 
   ///Gets the progress for the current page. The progress value is between 0 and 100.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#getProgress()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1415007-estimatedprogress
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getProgress](https://developer.android.com/reference/android/webkit/WebView#getProgress()))
+  ///- iOS ([Official API - WKWebView.estimatedProgress](https://developer.apple.com/documentation/webkit/wkwebview/1415007-estimatedprogress))
   Future<int?> getProgress() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('getProgress', args);
@@ -1138,21 +1153,23 @@ class InAppWebViewController {
 
   ///Loads the given [urlRequest].
   ///
-  ///[iosAllowingReadAccessTo], used in combination with [urlRequest] (using the `file://` scheme),
-  ///is an iOS-specific argument that represents the URL from which to read the web content.
+  ///- [allowingReadAccessTo], used in combination with [urlRequest] (using the `file://` scheme),
+  ///it represents the URL from which to read the web content.
   ///This URL must be a file-based URL (using the `file://` scheme).
   ///Specify the same value as the URL parameter to prevent WebView from reading any other content.
   ///Specify a directory to give WebView permission to read additional files in the specified directory.
+  ///**NOTE**: available only on iOS.
   ///
   ///**NOTE for Android**: when loading an URL Request using "POST" method, headers are ignored.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#loadUrl(java.lang.String)
-  ///
-  ///**Official iOS API**:
-  ///- https://developer.apple.com/documentation/webkit/wkwebview/1414954-load
-  ///- if [iosAllowingReadAccessTo] is used, https://developer.apple.com/documentation/webkit/wkwebview/1414973-loadfileurl
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.loadUrl](https://developer.android.com/reference/android/webkit/WebView#loadUrl(java.lang.String))). If method is "POST", [Official API - WebView.postUrl](https://developer.android.com/reference/android/webkit/WebView#postUrl(java.lang.String,%20byte[]))
+  ///- iOS ([Official API - WKWebView.load](https://developer.apple.com/documentation/webkit/wkwebview/1414954-load). If [allowingReadAccessTo] is used, [Official API - WKWebView.loadFileURL](https://developer.apple.com/documentation/webkit/wkwebview/1414973-loadfileurl))
   Future<void> loadUrl(
-      {required URLRequest urlRequest, Uri? iosAllowingReadAccessTo}) async {
+      {required URLRequest urlRequest,
+      @Deprecated('Use `allowingReadAccessTo` instead')
+          Uri? iosAllowingReadAccessTo,
+      Uri? allowingReadAccessTo}) async {
     assert(urlRequest.url != null && urlRequest.url.toString().isNotEmpty);
     assert(iosAllowingReadAccessTo == null ||
         iosAllowingReadAccessTo.isScheme("file"));
@@ -1160,19 +1177,24 @@ class InAppWebViewController {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('urlRequest', () => urlRequest.toMap());
     args.putIfAbsent(
-        'allowingReadAccessTo', () => iosAllowingReadAccessTo?.toString());
+        'allowingReadAccessTo',
+        () =>
+            allowingReadAccessTo?.toString() ??
+            iosAllowingReadAccessTo?.toString());
     await _channel.invokeMethod('loadUrl', args);
   }
 
   ///Loads the given [url] with [postData] (x-www-form-urlencoded) using `POST` method into this WebView.
-  ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#postUrl(java.lang.String,%20byte[])
   ///
   ///Example
   ///```dart
   ///var postData = Uint8List.fromList(utf8.encode("firstname=Foo&surname=Bar"));
   ///controller.postUrl(url: Uri.parse("https://www.example.com/"), postData: postData);
   ///```
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.postUrl](https://developer.android.com/reference/android/webkit/WebView#postUrl(java.lang.String,%20byte[])))
+  ///- iOS
   Future<void> postUrl({required Uri url, required Uint8List postData}) async {
     assert(url.toString().isNotEmpty);
     Map<String, dynamic> args = <String, dynamic>{};
@@ -1183,27 +1205,30 @@ class InAppWebViewController {
 
   ///Loads the given [data] into this WebView, using [baseUrl] as the base URL for the content.
   ///
-  ///- [mimeType] parameter specifies the format of the data. The default value is `"text/html"`.
-  ///- [encoding] parameter specifies the encoding of the data. The default value is `"utf8"`.
-  ///- [androidHistoryUrl] parameter is the URL to use as the history entry. The default value is `about:blank`. If non-null, this must be a valid URL. This parameter is used only on Android.
-  ///- [iosAllowingReadAccessTo], used in combination with [baseUrl] (using the `file://` scheme),
-  ///is an iOS-specific argument that represents the URL from which to read the web content.
+  ///- [mimeType] argument specifies the format of the data. The default value is `"text/html"`.
+  ///- [encoding] argument specifies the encoding of the data. The default value is `"utf8"`.
+  ///- [historyUrl] is an Android-specific argument that represents the URL to use as the history entry. The default value is `about:blank`. If non-null, this must be a valid URL.
+  ///- [allowingReadAccessTo], used in combination with [baseUrl] (using the `file://` scheme),
+  ///it represents the URL from which to read the web content.
   ///This [baseUrl] must be a file-based URL (using the `file://` scheme).
   ///Specify the same value as the [baseUrl] parameter to prevent WebView from reading any other content.
   ///Specify a directory to give WebView permission to read additional files in the specified directory.
+  ///**NOTE**: available only on iOS.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#loadDataWithBaseURL(java.lang.String,%20java.lang.String,%20java.lang.String,%20java.lang.String,%20java.lang.String)
-  ///
-  ///**Official iOS API**:
-  ///- https://developer.apple.com/documentation/webkit/wkwebview/1415004-loadhtmlstring
-  ///- https://developer.apple.com/documentation/webkit/wkwebview/1415011-load
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.loadDataWithBaseURL](https://developer.android.com/reference/android/webkit/WebView#loadDataWithBaseURL(java.lang.String,%20java.lang.String,%20java.lang.String,%20java.lang.String,%20java.lang.String)))
+  ///- iOS ([Official API - WKWebView.loadHTMLString](https://developer.apple.com/documentation/webkit/wkwebview/1415004-loadhtmlstring) or [Official API - WKWebView.load](https://developer.apple.com/documentation/webkit/wkwebview/1415011-load))
   Future<void> loadData(
       {required String data,
       String mimeType = "text/html",
       String encoding = "utf8",
       Uri? baseUrl,
-      Uri? androidHistoryUrl,
-      Uri? iosAllowingReadAccessTo}) async {
+      @Deprecated('Use `historyUrl` instead')
+          Uri? androidHistoryUrl,
+      Uri? historyUrl,
+      @Deprecated('Use `allowingReadAccessTo` instead')
+          Uri? iosAllowingReadAccessTo,
+      Uri? allowingReadAccessTo}) async {
     assert(iosAllowingReadAccessTo == null ||
         iosAllowingReadAccessTo.isScheme("file"));
 
@@ -1213,9 +1238,16 @@ class InAppWebViewController {
     args.putIfAbsent('encoding', () => encoding);
     args.putIfAbsent('baseUrl', () => baseUrl?.toString() ?? "about:blank");
     args.putIfAbsent(
-        'historyUrl', () => androidHistoryUrl?.toString() ?? "about:blank");
+        'historyUrl',
+        () =>
+            historyUrl?.toString() ??
+            androidHistoryUrl?.toString() ??
+            "about:blank");
     args.putIfAbsent(
-        'allowingReadAccessTo', () => iosAllowingReadAccessTo?.toString());
+        'allowingReadAccessTo',
+        () =>
+            allowingReadAccessTo?.toString() ??
+            iosAllowingReadAccessTo?.toString());
     await _channel.invokeMethod('loadData', args);
   }
 
@@ -1248,6 +1280,10 @@ class InAppWebViewController {
   ///controller.loadFile(assetFilePath: "assets/index.html");
   ///...
   ///```
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.loadUrl](https://developer.android.com/reference/android/webkit/WebView#loadUrl(java.lang.String)))
+  ///- iOS ([Official API - WKWebView.load](https://developer.apple.com/documentation/webkit/wkwebview/1414954-load))
   Future<void> loadFile({required String assetFilePath}) async {
     assert(assetFilePath.isNotEmpty);
     Map<String, dynamic> args = <String, dynamic>{};
@@ -1257,9 +1293,9 @@ class InAppWebViewController {
 
   ///Reloads the WebView.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#reload()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414969-reload
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.reload](https://developer.android.com/reference/android/webkit/WebView#reload()))
+  ///- iOS ([Official API - WKWebView.reload](https://developer.apple.com/documentation/webkit/wkwebview/1414969-reload))
   Future<void> reload() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('reload', args);
@@ -1267,9 +1303,9 @@ class InAppWebViewController {
 
   ///Goes back in the history of the WebView.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#goBack()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414952-goback
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.goBack](https://developer.android.com/reference/android/webkit/WebView#goBack()))
+  ///- iOS ([Official API - WKWebView.goBack](https://developer.apple.com/documentation/webkit/wkwebview/1414952-goback))
   Future<void> goBack() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('goBack', args);
@@ -1277,9 +1313,9 @@ class InAppWebViewController {
 
   ///Returns a boolean value indicating whether the WebView can move backward.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#canGoBack()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414966-cangoback
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.canGoBack](https://developer.android.com/reference/android/webkit/WebView#canGoBack()))
+  ///- iOS ([Official API - WKWebView.canGoBack](https://developer.apple.com/documentation/webkit/wkwebview/1414966-cangoback))
   Future<bool> canGoBack() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('canGoBack', args);
@@ -1287,9 +1323,9 @@ class InAppWebViewController {
 
   ///Goes forward in the history of the WebView.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#goForward()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414993-goforward
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.goForward](https://developer.android.com/reference/android/webkit/WebView#goForward()))
+  ///- iOS ([Official API - WKWebView.goForward](https://developer.apple.com/documentation/webkit/wkwebview/1414993-goforward))
   Future<void> goForward() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('goForward', args);
@@ -1297,9 +1333,9 @@ class InAppWebViewController {
 
   ///Returns a boolean value indicating whether the WebView can move forward.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#canGoForward()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414962-cangoforward
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.canGoForward](https://developer.android.com/reference/android/webkit/WebView#canGoForward()))
+  ///- iOS ([Official API - WKWebView.canGoForward](https://developer.apple.com/documentation/webkit/wkwebview/1414962-cangoforward))
   Future<bool> canGoForward() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('canGoForward', args);
@@ -1307,9 +1343,9 @@ class InAppWebViewController {
 
   ///Goes to the history item that is the number of steps away from the current item. Steps is negative if backward and positive if forward.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#goBackOrForward(int)
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414991-go
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.goBackOrForward](https://developer.android.com/reference/android/webkit/WebView#goBackOrForward(int)))
+  ///- iOS ([Official API - WKWebView.go](https://developer.apple.com/documentation/webkit/wkwebview/1414991-go))
   Future<void> goBackOrForward({required int steps}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('steps', () => steps);
@@ -1318,7 +1354,9 @@ class InAppWebViewController {
 
   ///Returns a boolean value indicating whether the WebView can go back or forward the given number of steps. Steps is negative if backward and positive if forward.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#canGoBackOrForward(int)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.canGoBackOrForward](https://developer.android.com/reference/android/webkit/WebView#canGoBackOrForward(int)))
+  ///- iOS
   Future<bool> canGoBackOrForward({required int steps}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('steps', () => steps);
@@ -1326,13 +1364,22 @@ class InAppWebViewController {
   }
 
   ///Navigates to a [WebHistoryItem] from the back-forward [WebHistory.list] and sets it as the current item.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> goTo({required WebHistoryItem historyItem}) async {
-    if (historyItem.offset != null) {
-      await goBackOrForward(steps: historyItem.offset!);
+    var steps = historyItem.offset;
+    if (steps != null) {
+      await goBackOrForward(steps: steps);
     }
   }
 
   ///Check if the WebView instance is in a loading state.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<bool> isLoading() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('isLoading', args);
@@ -1340,9 +1387,9 @@ class InAppWebViewController {
 
   ///Stops the WebView from loading.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#stopLoading()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414981-stoploading
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.stopLoading](https://developer.android.com/reference/android/webkit/WebView#stopLoading()))
+  ///- iOS ([Official API - WKWebView.stopLoading](https://developer.apple.com/documentation/webkit/wkwebview/1414981-stoploading))
   Future<void> stopLoading() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('stopLoading', args);
@@ -1362,19 +1409,22 @@ class InAppWebViewController {
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#evaluateJavascript(java.lang.String,%20android.webkit.ValueCallback%3Cjava.lang.String%3E)
-  ///
-  ///**Official iOS API**:
-  ///- https://developer.apple.com/documentation/webkit/wkwebview/1415017-evaluatejavascript
-  ///- https://developer.apple.com/documentation/webkit/wkwebview/3656442-evaluatejavascript
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.evaluateJavascript](https://developer.android.com/reference/android/webkit/WebView#evaluateJavascript(java.lang.String,%20android.webkit.ValueCallback%3Cjava.lang.String%3E)))
+  ///- iOS ([Official API - WKWebView.evaluateJavascript](https://developer.apple.com/documentation/webkit/wkwebview/3656442-evaluatejavascript))
   Future<dynamic> evaluateJavascript(
       {required String source, ContentWorld? contentWorld}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('source', () => source);
     args.putIfAbsent('contentWorld', () => contentWorld?.toMap());
     var data = await _channel.invokeMethod('evaluateJavascript', args);
-    if (data != null && defaultTargetPlatform == TargetPlatform.android)
-      data = json.decode(data);
+    if (data != null && defaultTargetPlatform == TargetPlatform.android) {
+      try {
+        // try to json decode the data coming from JavaScript
+        // otherwise return it as it is.
+        data = json.decode(data);
+      } catch (e) {}
+    }
     return data;
   }
 
@@ -1386,6 +1436,10 @@ class InAppWebViewController {
   ///because, in these events, the [WebView] is not ready to handle it yet.
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> injectJavascriptFileFromUrl(
       {required Uri urlFile,
       ScriptHtmlTagAttributes? scriptHtmlTagAttributes}) async {
@@ -1407,6 +1461,10 @@ class InAppWebViewController {
   ///because, in these events, the [WebView] is not ready to handle it yet.
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<dynamic> injectJavascriptFileFromAsset(
       {required String assetFilePath}) async {
     String source = await rootBundle.loadString(assetFilePath);
@@ -1419,6 +1477,10 @@ class InAppWebViewController {
   ///because, in these events, the [WebView] is not ready to handle it yet.
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> injectCSSCode({required String source}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('source', () => source);
@@ -1433,6 +1495,10 @@ class InAppWebViewController {
   ///because, in these events, the [WebView] is not ready to handle it yet.
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> injectCSSFileFromUrl(
       {required Uri urlFile,
       CSSLinkHtmlTagAttributes? cssLinkHtmlTagAttributes}) async {
@@ -1450,6 +1516,10 @@ class InAppWebViewController {
   ///because, in these events, the [WebView] is not ready to handle it yet.
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> injectCSSFileFromAsset({required String assetFilePath}) async {
     String source = await rootBundle.loadString(assetFilePath);
     await injectCSSCode(source: source);
@@ -1505,6 +1575,10 @@ class InAppWebViewController {
   ///**NOTE**: This method should be called, for example, in the [WebView.onWebViewCreated] or [WebView.onLoadStart] events or, at least,
   ///before you know that your JavaScript code will call the `window.flutter_inappwebview.callHandler` method,
   ///otherwise you won't be able to intercept the JavaScript message.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   void addJavaScriptHandler(
       {required String handlerName,
       required JavaScriptHandlerCallback callback}) {
@@ -1516,6 +1590,10 @@ class InAppWebViewController {
   ///Removes a JavaScript message handler previously added with the [addJavaScriptHandler()] associated to [handlerName] key.
   ///Returns the value associated with [handlerName] before it was removed.
   ///Returns `null` if [handlerName] was not found.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   JavaScriptHandlerCallback? removeJavaScriptHandler(
       {required String handlerName}) {
     return this.javaScriptHandlersMap.remove(handlerName);
@@ -1527,7 +1605,9 @@ class InAppWebViewController {
   ///
   ///**NOTE for iOS**: available on iOS 11.0+.
   ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/2873260-takesnapshot
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS ([Official API - WKWebView.takeSnapshot](https://developer.apple.com/documentation/webkit/wkwebview/2873260-takesnapshot))
   Future<Uint8List?> takeScreenshot(
       {ScreenshotConfiguration? screenshotConfiguration}) async {
     Map<String, dynamic> args = <String, dynamic>{};
@@ -1537,6 +1617,10 @@ class InAppWebViewController {
   }
 
   ///Sets the WebView options with the new [options] and evaluates them.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> setOptions({required InAppWebViewGroupOptions options}) async {
     Map<String, dynamic> args = <String, dynamic>{};
 
@@ -1545,6 +1629,10 @@ class InAppWebViewController {
   }
 
   ///Gets the current WebView options. Returns `null` if it wasn't able to get them.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<InAppWebViewGroupOptions?> getOptions() async {
     Map<String, dynamic> args = <String, dynamic>{};
 
@@ -1563,9 +1651,9 @@ class InAppWebViewController {
   ///Multiple calls to this method may return different objects.
   ///The object returned from this method will not be updated to reflect any new state.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#copyBackForwardList()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wkwebview/1414977-backforwardlist
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.copyBackForwardList](https://developer.android.com/reference/android/webkit/WebView#copyBackForwardList()))
+  ///- iOS ([Official API - WKWebView.backForwardList](https://developer.apple.com/documentation/webkit/wkwebview/1414977-backforwardlist))
   Future<WebHistory?> getCopyBackForwardList() async {
     Map<String, dynamic> args = <String, dynamic>{};
     Map<String, dynamic>? result =
@@ -1574,7 +1662,11 @@ class InAppWebViewController {
     return WebHistory.fromMap(result);
   }
 
-  ///Clears all the webview's cache.
+  ///Clears all the WebView's cache.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<void> clearCache() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('clearCache', args);
@@ -1584,24 +1676,28 @@ class InAppWebViewController {
   ///
   ///[find] represents the string to find.
   ///
-  ///**NOTE**: on Android, it finds all instances asynchronously. Successive calls to this will cancel any pending searches.
+  ///**NOTE**: on Android native WebView, it finds all instances asynchronously. Successive calls to this will cancel any pending searches.
   ///
   ///**NOTE**: on iOS, this is implemented using CSS and Javascript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#findAllAsync(java.lang.String)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.findAllAsync](https://developer.android.com/reference/android/webkit/WebView#findAllAsync(java.lang.String)))
+  ///- iOS
   Future<void> findAllAsync({required String find}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('find', () => find);
     await _channel.invokeMethod('findAllAsync', args);
   }
 
-  ///Highlights and scrolls to the next match found by [findAllAsync()]. Notifies [WebView.onFindResultReceived] listener.
+  ///Highlights and scrolls to the next match found by [findAllAsync]. Notifies [WebView.onFindResultReceived] listener.
   ///
   ///[forward] represents the direction to search.
   ///
   ///**NOTE**: on iOS, this is implemented using CSS and Javascript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#findNext(boolean)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.findNext](https://developer.android.com/reference/android/webkit/WebView#findNext(boolean)))
+  ///- iOS
   Future<void> findNext({required bool forward}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('forward', () => forward);
@@ -1612,22 +1708,32 @@ class InAppWebViewController {
   ///
   ///**NOTE**: on iOS, this is implemented using CSS and Javascript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#clearMatches()
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.clearMatches](https://developer.android.com/reference/android/webkit/WebView#clearMatches()))
+  ///- iOS
   Future<void> clearMatches() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('clearMatches', args);
   }
 
-  ///Gets the html (with javascript) of the Chromium's t-rex runner game. Used in combination with [getTRexRunnerCss()].
+  ///Gets the html (with javascript) of the Chromium's t-rex runner game. Used in combination with [getTRexRunnerCss].
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<String> getTRexRunnerHtml() async {
-    return await rootBundle
-        .loadString("packages/flutter_inappwebview/t_rex_runner/t-rex.html");
+    return await rootBundle.loadString(
+        "packages/flutter_inappwebview/assets/t_rex_runner/t-rex.html");
   }
 
-  ///Gets the css of the Chromium's t-rex runner game. Used in combination with [getTRexRunnerHtml()].
+  ///Gets the css of the Chromium's t-rex runner game. Used in combination with [getTRexRunnerHtml].
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<String> getTRexRunnerCss() async {
-    return await rootBundle
-        .loadString("packages/flutter_inappwebview/t_rex_runner/t-rex.css");
+    return await rootBundle.loadString(
+        "packages/flutter_inappwebview/assets/t_rex_runner/t-rex.css");
   }
 
   ///Scrolls the WebView to the position.
@@ -1638,9 +1744,9 @@ class InAppWebViewController {
   ///
   ///[animated] `true` to animate the scroll transition, `false` to make the scoll transition immediate.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/view/View#scrollTo(int,%20int)
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiscrollview/1619400-setcontentoffset
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - View.scrollTo](https://developer.android.com/reference/android/view/View#scrollTo(int,%20int)))
+  ///- iOS ([Official API - UIScrollView.setContentOffset](https://developer.apple.com/documentation/uikit/uiscrollview/1619400-setcontentoffset))
   Future<void> scrollTo(
       {required int x, required int y, bool animated = false}) async {
     Map<String, dynamic> args = <String, dynamic>{};
@@ -1658,9 +1764,9 @@ class InAppWebViewController {
   ///
   ///[animated] `true` to animate the scroll transition, `false` to make the scoll transition immediate.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/view/View#scrollBy(int,%20int)
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiscrollview/1619400-setcontentoffset
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - View.scrollBy](https://developer.android.com/reference/android/view/View#scrollBy(int,%20int)))
+  ///- iOS ([Official API - UIScrollView.setContentOffset](https://developer.apple.com/documentation/uikit/uiscrollview/1619400-setcontentoffset))
   Future<void> scrollBy(
       {required int x, required int y, bool animated = false}) async {
     Map<String, dynamic> args = <String, dynamic>{};
@@ -1670,12 +1776,14 @@ class InAppWebViewController {
     await _channel.invokeMethod('scrollBy', args);
   }
 
-  ///On Android, it pauses all layout, parsing, and JavaScript timers for all WebViews.
+  ///On Android native WebView, it pauses all layout, parsing, and JavaScript timers for all WebViews.
   ///This is a global requests, not restricted to just this WebView. This can be useful if the application has been paused.
   ///
-  ///On iOS, it is restricted to just this WebView.
+  ///On iOS, it is implemented using JavaScript and it is restricted to just this WebView.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#pauseTimers()
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.pauseTimers](https://developer.android.com/reference/android/webkit/WebView#pauseTimers()))
+  ///- iOS
   Future<void> pauseTimers() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('pauseTimers', args);
@@ -1683,9 +1791,11 @@ class InAppWebViewController {
 
   ///On Android, it resumes all layout, parsing, and JavaScript timers for all WebViews. This will resume dispatching all timers.
   ///
-  ///On iOS, it resumes all layout, parsing, and JavaScript timers to just this WebView.
+  ///On iOS, it is implemented using JavaScript and it resumes all layout, parsing, and JavaScript timers to just this WebView.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#resumeTimers()
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.resumeTimers](https://developer.android.com/reference/android/webkit/WebView#resumeTimers()))
+  ///- iOS
   Future<void> resumeTimers() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('resumeTimers', args);
@@ -1695,9 +1805,9 @@ class InAppWebViewController {
   ///
   ///**NOTE**: available on Android 21+.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/print/PrintManager
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiprintinteractioncontroller
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - PrintManager](https://developer.android.com/reference/android/print/PrintManager))
+  ///- iOS ([Official API - UIPrintInteractionController](https://developer.apple.com/documentation/uikit/uiprintinteractioncontroller))
   Future<void> printCurrentPage() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('printCurrentPage', args);
@@ -1705,9 +1815,9 @@ class InAppWebViewController {
 
   ///Gets the height of the HTML content.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#getContentHeight()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiscrollview/1619399-contentsize
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getContentHeight](https://developer.android.com/reference/android/webkit/WebView#getContentHeight()))
+  ///- iOS ([Official API - UIScrollView.contentSize](https://developer.apple.com/documentation/uikit/uiscrollview/1619399-contentsize))
   Future<int?> getContentHeight() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('getContentHeight', args);
@@ -1717,16 +1827,18 @@ class InAppWebViewController {
   ///
   ///[zoomFactor] represents the zoom factor to apply. On Android, the zoom factor will be clamped to the Webview's zoom limits and, also, this value must be in the range 0.01 (excluded) to 100.0 (included).
   ///
-  ///[iosAnimated] `true` to animate the transition to the new scale, `false` to make the transition immediate.
+  ///[animated] `true` to animate the transition to the new scale, `false` to make the transition immediate.
   ///**NOTE**: available only on iOS.
   ///
   ///**NOTE**: available on Android 21+.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#zoomBy(float)
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiscrollview/1619412-setzoomscale
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.zoomBy](https://developer.android.com/reference/android/webkit/WebView#zoomBy(float)))
+  ///- iOS ([Official API - UIScrollView.setZoomScale](https://developer.apple.com/documentation/uikit/uiscrollview/1619412-setzoomscale))
   Future<void> zoomBy(
-      {required double zoomFactor, bool iosAnimated = false}) async {
+      {required double zoomFactor,
+      @Deprecated('Use `animated` instead') bool? iosAnimated,
+      bool animated = false}) async {
     assert(defaultTargetPlatform != TargetPlatform.android ||
         (defaultTargetPlatform == TargetPlatform.android &&
             zoomFactor > 0.01 &&
@@ -1734,13 +1846,28 @@ class InAppWebViewController {
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('zoomFactor', () => zoomFactor);
-    args.putIfAbsent('iosAnimated', () => iosAnimated);
+    args.putIfAbsent('animated', () => iosAnimated ?? animated);
     return await _channel.invokeMethod('zoomBy', args);
+  }
+
+  ///Gets the URL that was originally requested for the current page.
+  ///This is not always the same as the URL passed to [InAppWebView.onLoadStarted] because although the load for that URL has begun,
+  ///the current page may not have changed. Also, there may have been redirects resulting in a different URL to that originally requested.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getOriginalUrl](https://developer.android.com/reference/android/webkit/WebView#getOriginalUrl()))
+  ///- iOS
+  Future<Uri?> getOriginalUrl() async {
+    Map<String, dynamic> args = <String, dynamic>{};
+    String? url = await _channel.invokeMethod('getOriginalUrl', args);
+    return url != null ? Uri.parse(url) : null;
   }
 
   ///Gets the current zoom scale of the WebView.
   ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiscrollview/1619419-zoomscale
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS ([Official API - UIScrollView.zoomScale](https://developer.apple.com/documentation/uikit/uiscrollview/1619419-zoomscale))
   Future<double?> getZoomScale() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('getZoomScale', args);
@@ -1754,9 +1881,13 @@ class InAppWebViewController {
 
   ///Gets the selected text.
   ///
-  ///**NOTE**: This method is implemented with using JavaScript.
+  ///**NOTE**: this method is implemented with using JavaScript.
   ///
-  ///**NOTE for Android**: available only on Android 19+.
+  ///**NOTE for Android native WebView**: available only on Android 19+.
+  ///
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
   Future<String?> getSelectedText() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('getSelectedText', args);
@@ -1764,9 +1895,11 @@ class InAppWebViewController {
 
   ///Gets the hit result for hitting an HTML elements.
   ///
-  ///**NOTE**: On iOS it is implemented using JavaScript.
+  ///**NOTE**: On iOS, it is implemented using JavaScript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#getHitTestResult()
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getHitTestResult](https://developer.android.com/reference/android/webkit/WebView#getHitTestResult()))
+  ///- iOS
   Future<InAppWebViewHitTestResult?> getHitTestResult() async {
     Map<String, dynamic> args = <String, dynamic>{};
     Map<dynamic, dynamic>? hitTestResultMap =
@@ -1785,11 +1918,11 @@ class InAppWebViewController {
     return InAppWebViewHitTestResult(type: type, extra: extra);
   }
 
-  ///Clears the current focus. It will clear also, for example, the current text selection.
+  ///Clears the current focus. On iOS and Android native WebView, it will clear also, for example, the current text selection.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/view/ViewGroup#clearFocus()
-  ///
-  ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiresponder/1621097-resignfirstresponder
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - ViewGroup.clearFocus](https://developer.android.com/reference/android/view/ViewGroup#clearFocus()))
+  ///- iOS ([Official API - UIResponder.resignFirstResponder](https://developer.apple.com/documentation/uikit/uiresponder/1621097-resignfirstresponder))
   Future<void> clearFocus() async {
     Map<String, dynamic> args = <String, dynamic>{};
     return await _channel.invokeMethod('clearFocus', args);
@@ -1805,9 +1938,11 @@ class InAppWebViewController {
 
   ///Requests the anchor or image element URL at the last tapped point.
   ///
-  ///**NOTE**: On iOS it is implemented using JavaScript.
+  ///**NOTE**: On iOS, it is implemented using JavaScript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#requestFocusNodeHref(android.os.Message)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.requestFocusNodeHref](https://developer.android.com/reference/android/webkit/WebView#requestFocusNodeHref(android.os.Message))).
+  ///- iOS.
   Future<RequestFocusNodeHrefResult?> requestFocusNodeHref() async {
     Map<String, dynamic> args = <String, dynamic>{};
     Map<dynamic, dynamic>? result =
@@ -1823,9 +1958,11 @@ class InAppWebViewController {
 
   ///Requests the URL of the image last touched by the user.
   ///
-  ///**NOTE**: On iOS it is implemented using JavaScript.
+  ///**NOTE**: On iOS, it is implemented using JavaScript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#requestImageRef(android.os.Message)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.requestImageRef](https://developer.android.com/reference/android/webkit/WebView#requestImageRef(android.os.Message))).
+  ///- iOS.
   Future<RequestImageRefResult?> requestImageRef() async {
     Map<String, dynamic> args = <String, dynamic>{};
     Map<dynamic, dynamic>? result =
@@ -1947,7 +2084,9 @@ class InAppWebViewController {
 
   ///Gets the SSL certificate for the main top-level page or null if there is no certificate (the site is not secure).
   ///
-  ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#getCertificate()
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebView.getCertificate](https://developer.android.com/reference/android/webkit/WebView#getCertificate())).
+  ///- iOS.
   Future<SslCertificate?> getCertificate() async {
     Map<String, dynamic> args = <String, dynamic>{};
     Map<String, dynamic>? sslCertificateMap =
@@ -2155,11 +2294,13 @@ class InAppWebViewController {
   ///
   ///This method should be called when the page is loaded, for example, when the [WebView.onLoadStop] is fired, otherwise the [WebMessageChannel] won't work.
   ///
-  ///**NOTE for Android**: This method should only be called if [AndroidWebViewFeature.isFeatureSupported] returns `true` for [AndroidWebViewFeature.CREATE_WEB_MESSAGE_CHANNEL].
+  ///**NOTE for Android native WebView**: This method should only be called if [AndroidWebViewFeature.isFeatureSupported] returns `true` for [AndroidWebViewFeature.CREATE_WEB_MESSAGE_CHANNEL].
   ///
-  ///**NOTE for iOS**: This is implemented using Javascript.
+  ///**NOTE**: On iOS, it is implemented using JavaScript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/androidx/webkit/WebViewCompat#createWebMessageChannel(android.webkit.WebView)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebViewCompat.createWebMessageChannel](https://developer.android.com/reference/androidx/webkit/WebViewCompat#createWebMessageChannel(android.webkit.WebView))).
+  ///- iOS.
   Future<WebMessageChannel?> createWebMessageChannel() async {
     Map<String, dynamic> args = <String, dynamic>{};
     Map<String, dynamic>? result =
@@ -2173,11 +2314,13 @@ class InAppWebViewController {
   ///
   ///A target origin can be set as a wildcard ("*"). However this is not recommended.
   ///
-  ///**NOTE for Android**: This method should only be called if [AndroidWebViewFeature.isFeatureSupported] returns `true` for [AndroidWebViewFeature.POST_WEB_MESSAGE].
+  ///**NOTE for Android native WebView**: This method should only be called if [AndroidWebViewFeature.isFeatureSupported] returns `true` for [AndroidWebViewFeature.POST_WEB_MESSAGE].
   ///
-  ///**NOTE for iOS**: This is implemented using Javascript.
+  ///**NOTE**: On iOS, it is implemented using JavaScript.
   ///
-  ///**Official Android API**: https://developer.android.com/reference/androidx/webkit/WebViewCompat#postWebMessage(android.webkit.WebView,%20androidx.webkit.WebMessageCompat,%20android.net.Uri)
+  ///**Supported Platforms/Implementations**:
+  ///- Android native WebView ([Official API - WebViewCompat.postWebMessage](https://developer.android.com/reference/androidx/webkit/WebViewCompat#postWebMessage(android.webkit.WebView,%20androidx.webkit.WebMessageCompat,%20android.net.Uri))).
+  ///- iOS.
   Future<void> postWebMessage(
       {required WebMessage message, Uri? targetOrigin}) async {
     if (targetOrigin == null) {
