@@ -1,5 +1,4 @@
 import 'dart:collection';
-import 'dart:io';
 import 'dart:typed_data';
 import 'dart:convert';
 import 'dart:ui';
@@ -1967,7 +1966,8 @@ class AndroidActionModeMenuItem {
         return AndroidActionModeMenuItem.values
             .firstWhere((element) => element.toValue() == value);
       } catch (e) {
-        return null;
+        // maybe coming from a Bitwise OR operator
+        return AndroidActionModeMenuItem._internal(value);
       }
     }
     return null;
@@ -1985,9 +1985,9 @@ class AndroidActionModeMenuItem {
       case 4:
         return "MENU_ITEM_PROCESS_TEXT";
       case 0:
-      default:
         return "MENU_ITEM_NONE";
     }
+    return _value.toString();
   }
 
   ///No menu items should be disabled.
@@ -2005,6 +2005,9 @@ class AndroidActionModeMenuItem {
       const AndroidActionModeMenuItem._internal(4);
 
   bool operator ==(value) => value == _value;
+
+  AndroidActionModeMenuItem operator |(AndroidActionModeMenuItem value) =>
+      AndroidActionModeMenuItem._internal(value.toValue() | _value);
 
   @override
   int get hashCode => _value.hashCode;
@@ -4052,7 +4055,7 @@ class NavigationAction {
 
   Map<String, dynamic> toMap() {
     return {
-      "request": request.toString(),
+      "request": request.toMap(),
       "isForMainFrame": isForMainFrame,
       "androidHasGesture": androidHasGesture,
       "androidIsRedirect": androidIsRedirect,
@@ -6325,7 +6328,7 @@ class IOSShouldAllowDeprecatedTLSAction {
 
 ///A URL load request that is independent of protocol or URL scheme.
 class URLRequest {
-  ///The URL of the request.
+  ///The URL of the request. Setting this to `null` will load `about:blank`.
   Uri? url;
 
   ///The HTTP request method.
@@ -6844,6 +6847,434 @@ class AndroidPullToRefreshSize {
 
   ///Large size.
   static const LARGE = const AndroidPullToRefreshSize._internal(0);
+
+  bool operator ==(value) => value == _value;
+
+  @override
+  int get hashCode => _value.hashCode;
+}
+
+///Class that represents the [WebView] native implementation to be used.
+class WebViewImplementation {
+  final int _value;
+
+  const WebViewImplementation._internal(this._value);
+
+  static final Set<WebViewImplementation> values =
+      [WebViewImplementation.NATIVE].toSet();
+
+  static WebViewImplementation? fromValue(int? value) {
+    if (value != null) {
+      try {
+        return WebViewImplementation.values
+            .firstWhere((element) => element.toValue() == value);
+      } catch (e) {
+        return null;
+      }
+    }
+    return null;
+  }
+
+  int toValue() => _value;
+
+  @override
+  String toString() {
+    switch (_value) {
+      case 0:
+      default:
+        return "NATIVE";
+    }
+  }
+
+  ///Default native implementation, such as `WKWebView` for iOS and `android.webkit.WebView` for Android.
+  static const NATIVE = const WebViewImplementation._internal(0);
+
+  bool operator ==(value) => value == _value;
+
+  @override
+  int get hashCode => _value.hashCode;
+}
+
+///Class representing a download request of the WebView used by the event [WebView.onDownloadStartRequest].
+class DownloadStartRequest {
+  ///The full url to the content that should be downloaded.
+  Uri url;
+
+  ///the user agent to be used for the download.
+  String? userAgent;
+
+  ///Content-disposition http header, if present.
+  String? contentDisposition;
+
+  ///The mimetype of the content reported by the server.
+  String? mimeType;
+
+  ///The file size reported by the server.
+  int contentLength;
+
+  ///A suggested filename to use if saving the resource to disk.
+  String? suggestedFilename;
+
+  ///The name of the text encoding of the receiver, or `null` if no text encoding was specified.
+  String? textEncodingName;
+
+  DownloadStartRequest(
+      {required this.url,
+      this.userAgent,
+      this.contentDisposition,
+      this.mimeType,
+      required this.contentLength,
+      this.suggestedFilename,
+      this.textEncodingName});
+
+  static DownloadStartRequest? fromMap(Map<String, dynamic>? map) {
+    if (map == null) {
+      return null;
+    }
+
+    return DownloadStartRequest(
+        url: Uri.parse(map["url"]),
+        userAgent: map["userAgent"],
+        contentDisposition: map["contentDisposition"],
+        mimeType: map["mimeType"],
+        contentLength: map["contentLength"],
+        suggestedFilename: map["suggestedFilename"],
+        textEncodingName: map["textEncodingName"]);
+  }
+
+  Map<String, dynamic> toMap() {
+    return {
+      "url": url.toString(),
+      "userAgent": userAgent,
+      "contentDisposition": contentDisposition,
+      "mimeType": mimeType,
+      "contentLength": contentLength,
+      "suggestedFilename": suggestedFilename,
+      "textEncodingName": textEncodingName
+    };
+  }
+
+  Map<String, dynamic> toJson() {
+    return this.toMap();
+  }
+
+  @override
+  String toString() {
+    return toMap().toString();
+  }
+}
+
+///Android-specific class representing the share state that should be applied to the custom tab.
+class CustomTabsShareState {
+  final int _value;
+
+  const CustomTabsShareState._internal(this._value);
+
+  static final Set<CustomTabsShareState> values = [
+    CustomTabsShareState.SHARE_STATE_DEFAULT,
+    CustomTabsShareState.SHARE_STATE_ON,
+    CustomTabsShareState.SHARE_STATE_OFF,
+  ].toSet();
+
+  static CustomTabsShareState? fromValue(int? value) {
+    if (value != null) {
+      try {
+        return CustomTabsShareState.values
+            .firstWhere((element) => element.toValue() == value);
+      } catch (e) {
+        return null;
+      }
+    }
+    return null;
+  }
+
+  int toValue() => _value;
+
+  @override
+  String toString() {
+    switch (_value) {
+      case 1:
+        return "SHARE_STATE_ON";
+      case 2:
+        return "SHARE_STATE_OFF";
+      case 0:
+      default:
+        return "SHARE_STATE_DEFAULT";
+    }
+  }
+
+  ///Applies the default share settings depending on the browser.
+  static const SHARE_STATE_DEFAULT = const CustomTabsShareState._internal(0);
+
+  ///Shows a share option in the tab.
+  static const SHARE_STATE_ON = const CustomTabsShareState._internal(1);
+
+  ///Explicitly does not show a share option in the tab.
+  static const SHARE_STATE_OFF = const CustomTabsShareState._internal(2);
+
+  bool operator ==(value) => value == _value;
+
+  @override
+  int get hashCode => _value.hashCode;
+}
+
+///Android-class that represents display mode of a Trusted Web Activity.
+abstract class TrustedWebActivityDisplayMode {
+  Map<String, dynamic> toMap() {
+    return {};
+  }
+
+  Map<String, dynamic> toJson() {
+    return this.toMap();
+  }
+
+  @override
+  String toString() {
+    return toMap().toString();
+  }
+}
+
+///Android-class that represents the default display mode of a Trusted Web Activity.
+///The system UI (status bar, navigation bar) is shown, and the browser toolbar is hidden while the user is on a verified origin.
+class TrustedWebActivityDefaultDisplayMode
+    implements TrustedWebActivityDisplayMode {
+  String _type = "DEFAULT_MODE";
+
+  Map<String, dynamic> toMap() {
+    return {"type": _type};
+  }
+
+  Map<String, dynamic> toJson() {
+    return this.toMap();
+  }
+
+  @override
+  String toString() {
+    return toMap().toString();
+  }
+}
+
+///Android-class that represents the default display mode of a Trusted Web Activity.
+///The system UI (status bar, navigation bar) is shown, and the browser toolbar is hidden while the user is on a verified origin.
+class TrustedWebActivityImmersiveDisplayMode
+    implements TrustedWebActivityDisplayMode {
+  ///Whether the Trusted Web Activity should be in sticky immersive mode.
+  bool isSticky;
+
+  ///The constant defining how to deal with display cutouts.
+  AndroidLayoutInDisplayCutoutMode layoutInDisplayCutoutMode;
+
+  String _type = "IMMERSIVE_MODE";
+
+  TrustedWebActivityImmersiveDisplayMode(
+      {required this.isSticky, required this.layoutInDisplayCutoutMode});
+
+  static TrustedWebActivityImmersiveDisplayMode? fromMap(
+      Map<String, dynamic>? map) {
+    if (map == null) {
+      return null;
+    }
+
+    return TrustedWebActivityImmersiveDisplayMode(
+        isSticky: map["isSticky"],
+        layoutInDisplayCutoutMode: map["layoutInDisplayCutoutMode"]);
+  }
+
+  Map<String, dynamic> toMap() {
+    return {
+      "isSticky": isSticky,
+      "layoutInDisplayCutoutMode": layoutInDisplayCutoutMode.toValue(),
+      "type": _type
+    };
+  }
+
+  Map<String, dynamic> toJson() {
+    return this.toMap();
+  }
+
+  @override
+  String toString() {
+    return toMap().toString();
+  }
+}
+
+///Android-specific class representing the share state that should be applied to the custom tab.
+///
+///**NOTE**: available on Android 28+.
+class AndroidLayoutInDisplayCutoutMode {
+  final int _value;
+
+  const AndroidLayoutInDisplayCutoutMode._internal(this._value);
+
+  static final Set<AndroidLayoutInDisplayCutoutMode> values = [
+    AndroidLayoutInDisplayCutoutMode.DEFAULT,
+    AndroidLayoutInDisplayCutoutMode.SHORT_EDGES,
+    AndroidLayoutInDisplayCutoutMode.NEVER,
+    AndroidLayoutInDisplayCutoutMode.ALWAYS
+  ].toSet();
+
+  static AndroidLayoutInDisplayCutoutMode? fromValue(int? value) {
+    if (value != null) {
+      try {
+        return AndroidLayoutInDisplayCutoutMode.values
+            .firstWhere((element) => element.toValue() == value);
+      } catch (e) {
+        return null;
+      }
+    }
+    return null;
+  }
+
+  int toValue() => _value;
+
+  @override
+  String toString() {
+    switch (_value) {
+      case 1:
+        return "SHORT_EDGES";
+      case 2:
+        return "NEVER";
+      case 3:
+        return "ALWAYS";
+      case 0:
+      default:
+        return "DEFAULT";
+    }
+  }
+
+  ///With this default setting, content renders into the cutout area when displayed in portrait mode, but content is letterboxed when displayed in landscape mode.
+  ///
+  ///**NOTE**: available on Android 28+.
+  static const DEFAULT = const AndroidLayoutInDisplayCutoutMode._internal(0);
+
+  ///Content renders into the cutout area in both portrait and landscape modes.
+  ///
+  ///**NOTE**: available on Android 28+.
+  static const SHORT_EDGES =
+      const AndroidLayoutInDisplayCutoutMode._internal(1);
+
+  ///Content never renders into the cutout area.
+  ///
+  ///**NOTE**: available on Android 28+.
+  static const NEVER = const AndroidLayoutInDisplayCutoutMode._internal(2);
+
+  ///The window is always allowed to extend into the DisplayCutout areas on the all edges of the screen.
+  ///
+  ///**NOTE**: available on Android 30+.
+  static const ALWAYS = const AndroidLayoutInDisplayCutoutMode._internal(3);
+
+  bool operator ==(value) => value == _value;
+
+  @override
+  int get hashCode => _value.hashCode;
+}
+
+/// Android-specific class representing Screen Orientation Lock type value of a Trusted Web Activity:
+/// https://www.w3.org/TR/screen-orientation/#screenorientation-interface
+class TrustedWebActivityScreenOrientation {
+  final int _value;
+
+  const TrustedWebActivityScreenOrientation._internal(this._value);
+
+  static final Set<TrustedWebActivityScreenOrientation> values = [
+    TrustedWebActivityScreenOrientation.DEFAULT,
+    TrustedWebActivityScreenOrientation.PORTRAIT_PRIMARY,
+    TrustedWebActivityScreenOrientation.PORTRAIT_SECONDARY,
+    TrustedWebActivityScreenOrientation.LANDSCAPE_PRIMARY,
+    TrustedWebActivityScreenOrientation.LANDSCAPE_SECONDARY,
+    TrustedWebActivityScreenOrientation.ANY,
+    TrustedWebActivityScreenOrientation.LANDSCAPE,
+    TrustedWebActivityScreenOrientation.PORTRAIT,
+    TrustedWebActivityScreenOrientation.NATURAL,
+  ].toSet();
+
+  static TrustedWebActivityScreenOrientation? fromValue(int? value) {
+    if (value != null) {
+      try {
+        return TrustedWebActivityScreenOrientation.values
+            .firstWhere((element) => element.toValue() == value);
+      } catch (e) {
+        return null;
+      }
+    }
+    return null;
+  }
+
+  int toValue() => _value;
+
+  @override
+  String toString() {
+    switch (_value) {
+      case 1:
+        return "PORTRAIT_PRIMARY";
+      case 2:
+        return "PORTRAIT_SECONDARY";
+      case 3:
+        return "LANDSCAPE_PRIMARY";
+      case 4:
+        return "LANDSCAPE_SECONDARY";
+      case 5:
+        return "ANY";
+      case 6:
+        return "LANDSCAPE";
+      case 7:
+        return "PORTRAIT";
+      case 8:
+        return "NATURAL";
+      case 0:
+      default:
+        return "DEFAULT";
+    }
+  }
+
+  /// The default screen orientation is the set of orientations to which the screen is locked when
+  /// there is no current orientation lock.
+  static const DEFAULT = const TrustedWebActivityScreenOrientation._internal(0);
+
+  ///  Portrait-primary is an orientation where the screen width is less than or equal to the
+  ///  screen height. If the device's natural orientation is portrait, then it is in
+  ///  portrait-primary when held in that position.
+  static const PORTRAIT_PRIMARY =
+      const TrustedWebActivityScreenOrientation._internal(1);
+
+  /// Portrait-secondary is an orientation where the screen width is less than or equal to the
+  /// screen height. If the device's natural orientation is portrait, then it is in
+  /// portrait-secondary when rotated 180° from its natural position.
+  static const PORTRAIT_SECONDARY =
+      const TrustedWebActivityScreenOrientation._internal(2);
+
+  /// Landscape-primary is an orientation where the screen width is greater than the screen height.
+  /// If the device's natural orientation is landscape, then it is in landscape-primary when held
+  /// in that position.
+  static const LANDSCAPE_PRIMARY =
+      const TrustedWebActivityScreenOrientation._internal(3);
+
+  /// Landscape-secondary is an orientation where the screen width is greater than the
+  /// screen height. If the device's natural orientation is landscape, it is in
+  /// landscape-secondary when rotated 180° from its natural orientation.
+  static const LANDSCAPE_SECONDARY =
+      const TrustedWebActivityScreenOrientation._internal(4);
+
+  /// Any is an orientation that means the screen can be locked to any one of portrait-primary,
+  /// portrait-secondary, landscape-primary and landscape-secondary.
+  static const ANY = const TrustedWebActivityScreenOrientation._internal(5);
+
+  /// Landscape is an orientation where the screen width is greater than the screen height and
+  /// depending on platform convention locking the screen to landscape can represent
+  /// landscape-primary, landscape-secondary or both.
+  static const LANDSCAPE =
+      const TrustedWebActivityScreenOrientation._internal(6);
+
+  /// Portrait is an orientation where the screen width is less than or equal to the screen height
+  /// and depending on platform convention locking the screen to portrait can represent
+  /// portrait-primary, portrait-secondary or both.
+  static const PORTRAIT =
+      const TrustedWebActivityScreenOrientation._internal(7);
+
+  /// Natural is an orientation that refers to either portrait-primary or landscape-primary
+  /// depending on the device's usual orientation. This orientation is usually provided by
+  /// the underlying operating system.
+  static const NATURAL = const TrustedWebActivityScreenOrientation._internal(8);
 
   bool operator ==(value) => value == _value;
 
